==> SHARE/1_Sociodemo_econom/1_2_mbirth.Rmd <==
---
title: "Month of Birth"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/mbirthDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/mbirthDS.R')
```


# Data process

## SHARE

### Wave 1 

#### Study-specific variable description

| **Name** | `mobirth` |
|----------|------------|
| **Label** | `month of birth` |
| **Table name** | `sharew1_rel2-6-0_cv_r_whitout_nointw1` |
| **Missings** | `-2 = refusal`<br/>`-1 = don't know` |
| **Categories** | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |
| **Description** |  |



```{r assign1, echo=F}
opal.assign.table.tibble(o, 'SHARE_w1_mbirth','SHARE.sharew1_rel2-6-0_cv_r_whitout_nointw1',
                         variables=list('mobirth'), missings = TRUE)
SHARE_w1_mbirth <- opal.execute(o,'SHARE_w1_mbirth')
```

```{r local1, echo=F}
vari <- SHARE_w1_mbirth$mobirth
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w1_mbirth, aes(x=factor(mobirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep "january" into 1`
* `keep "february" into 2`
* `keep "march" into 3`
* `keep "april" into 4`
* `keep "may" into 5`
* `keep "june" into 6`
* `keep "july" into 7`
* `keep "august" into 8`
* `keep "september" into 9`
* `keep "october" into 10`
* `keep "november" into 11`
* `keep "december" into 12`
* `keep "today" into 1`
* `recode 13 "today" with the month_of_interview`
* `recode -2 "refusal" into 997`
* `recode -1 "don't know" into 998`

**R script:**

```{r harmo1}
SHARE_ds_w1 <- tibble(id=SHARE_w1_mbirth$id)
SHARE_ds_w1$mbirth <- car::recode(SHARE_w1_mbirth$mobirth, "-2=997; -1=998;  NA=999")
SHARE_ds_w1$mbirth <- labelled(SHARE_ds_w1$mbirth, labels = c("january"=1, "february"=2, "march"=3, "april"=4, "may"=5, "june"=6, "july"=7, "august"=8, "september"=9, "october"=10, "november"=11, "december"=12, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1, echo=F}
vari <- SHARE_ds_w1$mbirth
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w1, aes(x=factor(mbirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")
```

#### Validation
```{r crosstabulation1, echo=F}
BeforeH <- table(SHARE_w1_mbirth$mobirth)[c(3:14,1,2)]
AfterH <- table(SHARE_ds_w1$mbirth)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "-2->997", "-1->998")
kable(C)
```





```{r importOPAL1, echo=F}
label(SHARE_ds_w1$mbirth) <- ds_label
mbirth <- SHARE_ds_w1
save(mbirth, file = "../RData/w1/mbirth.RData")
#opal.assign.data(o, 'SHARE_ds_w1', SHARE_ds_w1)
#opal.symbol_import(o,'SHARE_ds_w1', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w1 <- SHARE_ds_w1
SHARE_m_ds_w1$mbirth <- car::recode(SHARE_m_ds_w1$mbirth, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w1', SHARE_m_ds_w1)
#opal.symbol_import(o,'SHARE_m_ds_w1', project='_Athlos_Harm_Dataset')
```





### Wave 2 

#### Study-specific variable description

| **Name** | `mobirth` |
|----------|------------|
| **Label** | `month of birth` |
| **Table name** | `sharew2_rel2-6-0_cv_r_without_noint` |
| **Missings** | `-2 = refusal`<br/>`-1 = don't know` |
| **Categories** | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |
| **Description** |  |


```{r assign2, echo=F}
opal.assign.table.tibble(o, 'SHARE_w2_mbirth','SHARE.sharew2_rel2-6-0_cv_r_without_noint',
                         variables=list('mobirth'), missings = TRUE)
SHARE_w2_mbirth <- opal.execute(o,'SHARE_w2_mbirth')
```

```{r local2, echo=F}
vari <- SHARE_w2_mbirth$mobirth
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w2_mbirth, aes(x=factor(mobirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep "january" into 1`
* `keep "february" into 2`
* `keep "march" into 3`
* `keep "april" into 4`
* `keep "may" into 5`
* `keep "june" into 6`
* `keep "july" into 7`
* `keep "august" into 8`
* `keep "september" into 9`
* `keep "october" into 10`
* `keep "november" into 11`
* `keep "december" into 12`
* `keep "today" into 1`
* `recode 13 "today" with the month_of_interview`
* `recode -2 "refusal" into 997`
* `recode -1 "don't know" into 998`

**R script:**

```{r harmo2}
SHARE_ds_w2 <- tibble(id=SHARE_w2_mbirth$id)
SHARE_ds_w2$mbirth <- car::recode(SHARE_w2_mbirth$mobirth, "-2=997; -1=998;  NA=999")
SHARE_ds_w2$mbirth <- labelled(SHARE_ds_w2$mbirth, labels = c("january"=1, "february"=2, "march"=3, "april"=4, "may"=5, "june"=6, "july"=7, "august"=8, "september"=9, "october"=10, "november"=11, "december"=12, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2, echo=F}
vari <- SHARE_ds_w2$mbirth
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w2, aes(x=factor(mbirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")
```

#### Validation
```{r crosstabulation2, echo=F}
BeforeH <- c(table(SHARE_w2_mbirth$mobirth)[c(3:14,1,2)], sum(is.na(SHARE_w2_mbirth$mobirth)))
AfterH <- table(SHARE_ds_w2$mbirth)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "-2->997", "-1->998", "NA->999")
kable(C)
```





```{r importOPAL2, echo=F}
label(SHARE_ds_w2$mbirth) <- ds_label
mbirth <- SHARE_ds_w2
save(mbirth, file = "../RData/w2/mbirth.RData")
#opal.assign.data(o, 'SHARE_ds_w2', SHARE_ds_w2)
#opal.symbol_import(o,'SHARE_ds_w2', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w2 <- SHARE_ds_w2
SHARE_m_ds_w2$mbirth <- car::recode(SHARE_m_ds_w2$mbirth, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w2', SHARE_m_ds_w2)
#opal.symbol_import(o,'SHARE_m_ds_w2', project='_Athlos_Harm_Dataset')
```




### Wave 3 

#### Study-specific variable description

| **Name** | `mobirth` |
|----------|------------|
| **Label** | `month of birth` |
| **Table name** | `sharew3_rel1_cv_r_without_noint` |
| **Missings** | `-2 = refusal`<br/>`-1 = don't know` |
| **Categories** | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |
| **Description** |  |


```{r assign3, echo=F}
opal.assign.table.tibble(o, 'SHARE_w3_mbirth','SHARE.sharew3_rel1_cv_r_without_noint',
                         variables=list('mobirth'), missings = TRUE)
SHARE_w3_mbirth <- opal.execute(o,'SHARE_w3_mbirth')
```

```{r local3, echo=F}
vari <- SHARE_w3_mbirth$mobirth
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w3_mbirth, aes(x=factor(mobirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep "january" into 1`
* `keep "february" into 2`
* `keep "march" into 3`
* `keep "april" into 4`
* `keep "may" into 5`
* `keep "june" into 6`
* `keep "july" into 7`
* `keep "august" into 8`
* `keep "september" into 9`
* `keep "october" into 10`
* `keep "november" into 11`
* `keep "december" into 12`
* `keep "today" into 1`
* `recode 13 "today" with the month_of_interview`
* `recode -2 "refusal" into 997`
* `recode -1 "don't know" into 998`

**R script:**

```{r harmo3}
SHARE_ds_w3 <- tibble(id=SHARE_w3_mbirth$id)
SHARE_ds_w3$mbirth <- car::recode(SHARE_w3_mbirth$mobirth, "-2=997; -1=998;  NA=999")
SHARE_ds_w3$mbirth <- labelled(SHARE_ds_w3$mbirth, labels = c("january"=1, "february"=2, "march"=3, "april"=4, "may"=5, "june"=6, "july"=7, "august"=8, "september"=9, "october"=10, "november"=11, "december"=12, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript3, echo=F}
vari <- SHARE_ds_w3$mbirth
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w3, aes(x=factor(mbirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")
```

#### Validation
```{r crosstabulation3, echo=F}
BeforeH <- c( table(SHARE_w3_mbirth$mobirth), sum(is.na(SHARE_w3_mbirth$mobirth)) )
AfterH <- table(SHARE_ds_w3$mbirth)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "NA->999")
kable(C)
```





```{r importOPAL3, echo=F}
label(SHARE_ds_w3$mbirth) <- ds_label
mbirth <- SHARE_ds_w3
save(mbirth, file = "../RData/w3/mbirth.RData")
#opal.assign.data(o, 'SHARE_ds_w3', SHARE_ds_w3)
#opal.symbol_import(o,'SHARE_ds_w3', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w3 <- SHARE_ds_w3
SHARE_m_ds_w3$mbirth <- car::recode(SHARE_m_ds_w3$mbirth, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w3', SHARE_m_ds_w3)
#opal.symbol_import(o,'SHARE_m_ds_w3', project='_Athlos_Harm_Dataset')
```






### Wave 4 

#### Study-specific variable description

| **Name** | `mobirth` |
|----------|------------|
| **Label** | `month of birth` |
| **Table name** | `sharew4_rel1-1-1_cv_r_without_noint` |
| **Missings** | `-2 = refusal`<br/>`-1 = don't know` |
| **Categories** | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |
| **Description** |  |



```{r assign4, echo=F}
opal.assign.table.tibble(o, 'SHARE_w4_mbirth','SHARE.sharew4_rel1-1-1_cv_r_without_noint',
                         variables=list('mobirth'), missings = TRUE)
SHARE_w4_mbirth <- opal.execute(o,'SHARE_w4_mbirth')
```

```{r local4, echo=F}
vari <- SHARE_w4_mbirth$mobirth
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w4_mbirth, aes(x=factor(mobirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep "january" into 1`
* `keep "february" into 2`
* `keep "march" into 3`
* `keep "april" into 4`
* `keep "may" into 5`
* `keep "june" into 6`
* `keep "july" into 7`
* `keep "august" into 8`
* `keep "september" into 9`
* `keep "october" into 10`
* `keep "november" into 11`
* `keep "december" into 12`
* `keep "today" into 1`
* `recode 13 "today" with the month_of_interview`
* `recode -2 "refusal" into 997`
* `recode -1 "don't know" into 998`

**R script:**

```{r harmo4}
SHARE_ds_w4 <- tibble(id=SHARE_w4_mbirth$id)
SHARE_ds_w4$mbirth <- car::recode(SHARE_w4_mbirth$mobirth, "-2=997; -1=998;  NA=999")
SHARE_ds_w4$mbirth <- labelled(SHARE_ds_w4$mbirth, labels = c("january"=1, "february"=2, "march"=3, "april"=4, "may"=5, "june"=6, "july"=7, "august"=8, "september"=9, "october"=10, "november"=11, "december"=12, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript4, echo=F}
vari <- SHARE_ds_w4$mbirth 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w4, aes(x=factor(mbirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")
```

#### Validation
```{r crosstabulation4, echo=F}
BeforeH <- c(table(SHARE_w4_mbirth$mobirth), sum(is.na(SHARE_w4_mbirth$mobirth)))
AfterH <- table(SHARE_ds_w4$mbirth)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "NA->999")
kable(C)
```





```{r importOPAL4, echo=F}
label(SHARE_ds_w4$mbirth) <- ds_label
mbirth <- SHARE_ds_w4
save(mbirth, file = "../RData/w4/mbirth.RData")
#opal.assign.data(o, 'SHARE_ds_w4', SHARE_ds_w4)
#opal.symbol_import(o,'SHARE_ds_w4', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w4 <- SHARE_ds_w4
SHARE_m_ds_w4$mbirth <- car::recode(SHARE_m_ds_w4$mbirth, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w4', SHARE_m_ds_w4)
#opal.symbol_import(o,'SHARE_m_ds_w4', project='_Athlos_Harm_Dataset')
```







### Wave 5 

#### Study-specific variable description

| **Name** | `mobirth` |
|----------|------------|
| **Label** | `month of birth` |
| **Table name** | `sharew5_rel1-0-0_cv_r_without_noint` |
| **Missings** | `-2 = refusal`<br/>`-1 = don't know` |
| **Categories** | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |
| **Description** |  |


```{r assign5, echo=F}
opal.assign.table.tibble(o, 'SHARE_w5_mbirth','SHARE.sharew5_rel1-0-0_cv_r_without_noint',
                         variables=list('mobirth'), missings = TRUE)
SHARE_w5_mbirth <- opal.execute(o,'SHARE_w5_mbirth')
```

```{r local5, echo=F}
vari <- SHARE_w5_mbirth$mobirth
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w5_mbirth, aes(x=factor(mobirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep "january" into 1`
* `keep "february" into 2`
* `keep "march" into 3`
* `keep "april" into 4`
* `keep "may" into 5`
* `keep "june" into 6`
* `keep "july" into 7`
* `keep "august" into 8`
* `keep "september" into 9`
* `keep "october" into 10`
* `keep "november" into 11`
* `keep "december" into 12`
* `keep "today" into 1`
* `recode 13 "today" with the month_of_interview`
* `recode -2 "refusal" into 997`
* `recode -1 "don't know" into 998`

**R script:**

```{r harmo5}
SHARE_ds_w5 <- tibble(id=SHARE_w5_mbirth$id)
SHARE_ds_w5$mbirth <- car::recode(SHARE_w5_mbirth$mobirth, "-2=997; -1=998;  NA=999")
SHARE_ds_w5$mbirth <- labelled(SHARE_ds_w5$mbirth, labels = c("january"=1, "february"=2, "march"=3, "april"=4, "may"=5, "june"=6, "july"=7, "august"=8, "september"=9, "october"=10, "november"=11, "december"=12, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript5, echo=F}
vari <- SHARE_ds_w5$mbirth
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w5, aes(x=factor(mbirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")
```

#### Validation
```{r crosstabulation5, echo=F}
BeforeH <- c(table(SHARE_w5_mbirth$mobirth), sum(is.na(SHARE_w5_mbirth$mobirth)))
AfterH <- table(SHARE_ds_w5$mbirth)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "NA->999")
kable(C)
```





```{r importOPAL5, echo=F}
label(SHARE_ds_w5$mbirth) <- ds_label
mbirth <- SHARE_ds_w5
save(mbirth, file = "../RData/w5/mbirth.RData")
#opal.assign.data(o, 'SHARE_ds_w5', SHARE_ds_w5)
#opal.symbol_import(o,'SHARE_ds_w5', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w5 <- SHARE_ds_w5
SHARE_m_ds_w5$mbirth <- car::recode(SHARE_m_ds_w5$mbirth, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w5', SHARE_m_ds_w5)
#opal.symbol_import(o,'SHARE_m_ds_w5', project='_Athlos_Harm_Dataset')
```


## Summary of descriptive statistics of the harmonised variable accross populations and waves





```{r summ, echo=F}

SHARE.cw <- list(w1 = "w1", w2 = "w2", w3 = "w3", w4 = "w4", w5 = "w5")

# Consider only harmonised waves
l.hds <- lapply(SHARE.cw, function(wname) if(exists(paste0("SHARE_ds_",wname))){wname = get(paste0("SHARE_ds_",wname))})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

# Labelling of the tibbles and creating new tibbles with all missings recodified as NA

for(name in names(l.hds)) {
  aux_object <- l.hds[[name]]
  label(l.hds[[name]][[2]]) <- label(aux_object[[2]]) <- ds_label
  l.hds[[name]][[2]] <- labelled(l.hds[[name]][[2]], labels = cont_label)
  aux_object[[2]] <- car::recode(aux_object[[2]], "miss_values_vector=NA")
  aux_object[[2]] <- labelled(aux_object[[2]], labels = cont_label)
  m_name <- paste0("SHARE_m_ds_",name)
  assign(m_name, aux_object)
  m.hds[[name]] <- get(paste0("SHARE_m_ds_",name))
  rm(aux_object)
}

t.summ <- summary(m.hds[[1]][2])[1:6]
t.hds <- substr(t.summ,1,regexpr(":", t.summ, fixed=T)-1)
t.hds <- c(substr(t.summ,1,regexpr(":", t.summ, fixed=T)-1),labels(Continuous_summary(l.hds[[1]][[2]],missing_values = miss_values_vector)$values_table)[[2]])
for (i in seq_along(l.hds)){
  t.summ <- summary(m.hds[[i]][2])[1:6]
  t.hds <- cbind(t.hds,c(as.numeric(substr(t.summ,regexpr(":", t.summ, fixed=T)+1,nchar(t.summ))),as.numeric(Continuous_summary(l.hds[[i]][[2]],missing_values = miss_values_vector)$values_table[1,])))
}
t.hds <- rbind(t.hds,c("sample size", sapply(l.hds,function(wave) length(wave[[1]]))))

dimnames(t.hds)[[2]] <- c(dimnames(summary(m.hds[[1]][2]))[[2]],toupper(names(l.hds)))
kable(t.hds,caption=ds_label)



```




# Quality estimation
No comments.


```{r closeRsession, echo=FALSE}
opal.logout(o)
```
