==> SHARE/1_Sociodemo_econom/2_age.Rmd <==
---
title: "Age"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/ageDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/ageDS.R')
id_missings <- c("Cf-397794-01", "DE-271503-01", "Ih-306138-02", "Ih-737574-01", "Ih-907477-01", "NL-027316-01", "NL-037198-01", "NL-073637-01", "NL-155292-01", "NL-372323-01", "NL-922959-01", "NL-924571-01", "GR-190740-02", "Bn-330112-02", "DK-639322-02", "ES-932056-02", "PL-691802-02", "SE-764326-05", "AT-215568-02", "AT-864946-01", "AT-886095-01", "CZ-023045-01", "CZ-033227-01", "CZ-191583-01", "CZ-356553-01", "CZ-373473-04", "CZ-384175-01", "CZ-526053-01", "CZ-638063-01", "CZ-694720-01", "CZ-695635-01", "F1-191516-01", "F1-728117-01", "HU-042621-01", "PT-022802-01", "PT-073544-01", "PT-443942-01", "PT-713628-01", "PT-843462-01", "DK-197676-02", "Eg-036951-05", "Eg-197395-02", "Eg-758727-02", "Eg-904129-02", "Ih-826469-03", "LU-989099-01", "NL-132447-01", "NL-756812-02", "NL-785371-02", "NL-948960-02", "SE-368759-02", "SE-607164-02", "SE-684973-02", "SI-139484-02")

```


# Data process

**Functions:**
```{r functions}



#function to compute the age
comp_age <- function(y_birth, m_birth, y_int, m_int, miss = miss_values_vector){
  N <- length(y_birth)
  #first, it's computed the age of all
  age <- rep(NA,N)
  age[which(m_int<m_birth)] <- (y_int[which(m_int<m_birth)]-y_birth[which(m_int<m_birth)])-1
  age[which(m_int>=m_birth)] <- y_int[which(m_int>=m_birth)]-y_birth[which(m_int>=m_birth)]
  age[which(m_birth>990)] <- y_int[which(m_birth>990)]-y_birth[which(m_birth>990)]
  
  #It's emphasized the individues having missing
  
  mis <- rep(0,N)
  for(i in 1:length(miss)){
    mis[which(y_birth==miss[i])] <- miss[i]
  }
  age[which(mis!=0)] <- mis[which(mis!=0)]

  mis <- rep(0,N)
  for(i in 1:length(miss)){
    mis[which(y_int==miss[i])] <- miss[i]
  }
  age[which(mis!=0)] <- mis[which(mis!=0)]
  
  
  age

}



```


## SHARE

### Wave 1 

#### Study-specific variable description

| **Name** | `yrbirth`|`mobirth`| `int_year` | `int_month` |
|-|-|-|-|-|
| **Label** | `year of birth`| `month of birth` | `interview year` | `interview month` |
| **Table name** | `sharew1_rel2-6-0_cv_r_whitout_nointw1` | `idem` | `idem` | `idem` |
| **Categories** |  | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |  | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |
| **Missings** | `-2 = refusal`<br/>`-1 = don't know` | `-2 = refusal`<br/>`-1 = don't know` |  | `-2 = refusal`<br/>`-1 = don't know` |
| **Description** |  |  |  |  |






```{r assign1, echo=F}
opal.assign.table.tibble(o, 'SHARE_w1','SHARE.sharew1_rel2-6-0_cv_r_whitout_nointw1',
                         variables=list('yrbirth', 'mobirth', 'int_year', 'int_month'), missings = TRUE)
SHARE_w1 <- opal.execute(o,'SHARE_w1')
```

```{r local1, echo=F}

vari <- SHARE_w1$yrbirth
vbl_miss <- c(-2,-1)
kable(Continuous_summary(var = vari, missing_values = c(-2,-1))[3], caption = "Year of birth")
kable(Continuous_summary(vari, missing_values = c(-2,-1))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-2,-1))$summary, caption = "Summary")
ggplot(SHARE_w1, aes(yrbirth)) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% vbl_miss)) + xlab("Year of birth") + ylab("Frequency")

vari <- SHARE_w1$mobirth
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "Month of birth") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w1, aes(x=factor(mobirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")

vari <- SHARE_w1$int_year
vbl_miss <- c(-2,-1)
kable(Continuous_summary(vari, missing_values = vbl_miss)[3], caption = "Year of interview") 
pander(Continuous_summary(vari, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SHARE_w1, aes(x=factor(int_year))) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% vbl_miss)) + xlab("Year of interview") + ylab("Frequency")

vari <- SHARE_w1$int_month
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "Month of interview") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w1, aes(x=factor(int_month))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of interview") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variables it has to be computed by applying the following formula: 

* `date of interview - date of birth`

**R script:**

```{r harmo1}
SHARE_ds_w1 <- tibble(id=SHARE_w1$id)


#harmonise year of birth

yrbirth_missing <- (SHARE_w1 %>% filter(is.na(yrbirth) | yrbirth %in% c(-1,-2)))$id
SHARE_miss <- SHARE_w1 %>% select(id,yrbirth) %>% filter(id %in% id_missings)
# Wave 1 missing yrbirth data is non-missing in posterior waves tables for some id's (look at validation of yrbirth at the end). Here is imputed.
SHARE_w1$yrbirth[which(SHARE_w1$id=="Ih-306138-02")] <- 1934
SHARE_w1$yrbirth[which(SHARE_w1$id=="NL-037198-01")] <- 1919

SHARE_w1$year_of_birth <- car::recode(SHARE_w1$yrbirth, "-2=997; -1=998;  NA=999")


#harmonise month of birth
SHARE_w1$month_of_birth <- car::recode(SHARE_w1$mobirth, "-2=997; -1=998;  NA=999")

#harmonise month of interview
SHARE_w1$month_of_interview <- car::recode(SHARE_w1$int_month, "-2=997; -1=998;  c(13,NA)=999")

#harmonise year of interview
SHARE_w1$year_of_interview <- car::recode(SHARE_w1$int_year, "NA=999")

SHARE_ds_w1$age <- comp_age(y_birth=SHARE_w1$year_of_birth , m_birth=SHARE_w1$month_of_birth ,
         y_int=SHARE_w1$year_of_interview , m_int=SHARE_w1$month_of_interview , miss=miss_values_vector)



```

#### Statistics of the new harmonised variable
```{r descript1, echo=F}
vari <- SHARE_ds_w1$age
kable(Continuous_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SHARE_ds_w1, aes(age)) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% miss_values_vector)) + xlab("Age") + ylab("Frequency")
```

#### Validation
```{r crosstabulation1, echo=F}
```






### Wave 2 

#### Study-specific variable description




| **Name** |`mobirth`| `yrbirth`| `int_month_w2` | `int_year_w2` |
|-|-|-|-|-|
| **Label** | `month of birth` | `year of birth`| `interview month` | `interview year` |
| **Table name** | `sharew2_rel2-6-0_cv_r_without_noint` | `idem` | `idem` | `idem` |
| **Categories** | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |  | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |  |
| **Missings** | `-2 = refusal`<br/>`-1 = don't know` | `-2 = refusal`<br/>`-1 = don't know` | `-2 = refusal`<br/>`-1 = don't know` |  |
| **Description** |  |  |  |  |

```{r assign2, echo=F}
opal.assign.table.tibble(o, 'SHARE_w2','SHARE.sharew2_rel2-6-0_cv_r_without_noint',
                         variables=list('mobirth', 'yrbirth', 'int_month_w2', 'int_year_w2'), missings = TRUE)
SHARE_w2 <- opal.execute(o,'SHARE_w2')
```

```{r local2, echo=F}
vari <- SHARE_w2$yrbirth
vbl_miss <- c(-2,-1)
kable(Continuous_summary(var = vari, missing_values = c(-2,-1))[3], caption = "Year of birth")
kable(Continuous_summary(vari, missing_values = c(-2,-1))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-2,-1))$summary, caption = "Summary")
ggplot(SHARE_w2, aes(yrbirth)) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% vbl_miss)) + xlab("Year of birth") + ylab("Frequency")

vari <- SHARE_w2$mobirth
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "Month of birth") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w2, aes(x=factor(mobirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")

vari <- SHARE_w2$int_year_w2
vbl_miss <- c(-2,-1)
kable(Continuous_summary(vari, missing_values = vbl_miss)[3], caption = "Year of interview") 
pander(Continuous_summary(vari, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SHARE_w2, aes(x=factor(int_year_w2))) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% vbl_miss)) + xlab("Year of interview") + ylab("Frequency")

vari <- SHARE_w2$int_month_w2
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "Month of interview") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w2, aes(x=factor(int_month_w2))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of interview") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variables it has to be computed by applying the following formula: 

* `date of interview - date of birth`

**R script:**

```{r harmo2}
SHARE_ds_w2 <- tibble(id=SHARE_w2$id)


#harmonise year of birth

#Some missing yrbirth data is available from previous waves
SHARE_w2 <- left_join(SHARE_w2,SHARE_w1 %>% select(id,yrbirth) %>% filter(!yrbirth %in% c(-1,-2)),by = "id", suffix = c("","_w1"))
SHARE_w2$yrbirth[which(is.na(SHARE_w2$yrbirth) | (SHARE_w2$yrbirth %in% c(-1,-2) & !is.na(SHARE_w2$yrbirth_w1)))] <- SHARE_w2$yrbirth_w1[which(is.na(SHARE_w2$yrbirth) | (SHARE_w2$yrbirth %in% c(-1,-2) & !is.na(SHARE_w2$yrbirth_w1)))]
SHARE_w2 <- SHARE_w2 %>% select(-yrbirth_w1)
yrbirth_missing <- union(yrbirth_missing,(SHARE_w2 %>% filter(is.na(yrbirth) | yrbirth %in% c(-1,-2)))$id)
SHARE_miss <- full_join(SHARE_miss,SHARE_w2 %>% select(id,yrbirth) %>% filter(id %in% id_missings),by="id",suffix = c("_w1","_w2"))

SHARE_w2$year_of_birth <- car::recode(SHARE_w2$yrbirth, "-2=997; -1=998;  NA=999")


#harmonise month of birth
SHARE_w2$month_of_birth <- car::recode(SHARE_w2$mobirth, "-2=997; -1=998;  NA=999")

#harmonise month of interview
SHARE_w2$month_of_interview <- car::recode(SHARE_w2$int_month_w2, "-2=997; -1=998;  c(13,NA)=999")

#harmonise year of interview
SHARE_w2$year_of_interview <- car::recode(SHARE_w2$int_year_w2, "NA=999")



SHARE_ds_w2$age <- comp_age(y_birth=SHARE_w2$year_of_birth , m_birth=SHARE_w2$month_of_birth ,
         y_int=SHARE_w2$year_of_interview , m_int=SHARE_w2$month_of_interview , miss=miss_values_vector)

```

#### Statistics of the new harmonised variable
```{r descript2, echo=F}
vari <- SHARE_ds_w2$age
kable(Continuous_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SHARE_ds_w2, aes(age)) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% miss_values_vector)) + xlab("Age") + ylab("Frequency")
```

#### Validation
```{r crosstabulation2, echo=F}
```






### Wave 3 

#### Study-specific variable description



| **Name** |`mobirth`| `yrbirth`| `int_month_w3` | `int_year_w3` |
|-|-|-|-|-|
| **Label** | `month of birth` | `year of birth`| `interview month` | `interview year` |
| **Table name** | `sharew3_rel1_cv_r_without_noint` | `idem` | `idem` | `idem` |
| **Categories** | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |  | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |  |
| **Missings** | `-2 = refusal`<br/>`-1 = don't know` | `NA` | `-2 = refusal`<br/>`-1 = don't know` |  |
| **Description** |  |  |  |  |



```{r assign3, echo=F}
opal.assign.table.tibble(o, 'SHARE_w3','SHARE.sharew3_rel1_cv_r_without_noint',
                         variables=list('yrbirth', 'mobirth', 'int_month_w3', 'int_year_w3'), missings = TRUE)
SHARE_w3 <- opal.execute(o,'SHARE_w3')
```

```{r local3, echo=F}
vari <- SHARE_w3$yrbirth
vbl_miss <- c(-2,-1)
kable(Continuous_summary(var = vari, missing_values = c(-2,-1))[3], caption = "Year of birth")
kable(Continuous_summary(vari, missing_values = c(-2,-1))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-2,-1))$summary, caption = "Summary")
ggplot(SHARE_w3, aes(yrbirth)) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% vbl_miss)) + xlab("Year of birth") + ylab("Frequency")

vari <- SHARE_w3$mobirth
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "Month of birth") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w3, aes(x=factor(mobirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")

vari <- SHARE_w3$int_year_w3
vbl_miss <- c(-2,-1)
kable(Continuous_summary(vari, missing_values = vbl_miss)[3], caption = "Year of interview") 
pander(Continuous_summary(vari, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SHARE_w3, aes(x=factor(int_year_w3))) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% vbl_miss)) + xlab("Year of interview") + ylab("Frequency")

vari <- SHARE_w3$int_month_w3
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "Month of interview") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w3, aes(x=factor(int_month_w3))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of interview") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variables it has to be computed by applying the following formula: 

* `date of interview - date of birth`

**R script:**

```{r harmo3}
SHARE_ds_w3 <- tibble(id=SHARE_w3$id)


#harmonise year of birth

#Some missing yrbirth data is available from previous waves
SHARE_w3 <- left_join(SHARE_w3,SHARE_w1 %>% select(id,yrbirth) %>% filter(!yrbirth %in% c(-1,-2)),by = "id", suffix = c("","_w1"))
SHARE_w3$yrbirth[which(is.na(SHARE_w3$yrbirth) | (SHARE_w3$yrbirth %in% c(-1,-2) & !is.na(SHARE_w3$yrbirth_w1)))] <- SHARE_w3$yrbirth_w1[which(is.na(SHARE_w3$yrbirth) | (SHARE_w3$yrbirth %in% c(-1,-2) & !is.na(SHARE_w3$yrbirth_w1)))]
SHARE_w3 <- left_join(SHARE_w3,SHARE_w2 %>% select(id,yrbirth) %>% filter(!yrbirth %in% c(-1,-2)),by = "id", suffix = c("","_w2"))
SHARE_w3$yrbirth[which(is.na(SHARE_w3$yrbirth) | (SHARE_w3$yrbirth %in% c(-1,-2) & !is.na(SHARE_w3$yrbirth_w2)))] <- SHARE_w3$yrbirth_w2[which(is.na(SHARE_w3$yrbirth) | (SHARE_w3$yrbirth %in% c(-1,-2) & !is.na(SHARE_w3$yrbirth_w2)))]
SHARE_w3 <- SHARE_w3 %>% select(-yrbirth_w1,-yrbirth_w2)
yrbirth_missing <- union(yrbirth_missing,(SHARE_w3 %>% filter(is.na(yrbirth) | yrbirth %in% c(-1,-2)))$id)
SHARE_miss <- full_join(SHARE_miss,SHARE_w3 %>% select(id,yrbirth) %>% filter(id %in% id_missings),by="id")
# Wave 3 missing yrbirth data is non-missing in posterior waves tables for some id's (look at validation of yrbirth at the end). Here is imputed.
SHARE_w3$yrbirth[SHARE_w3$id=="SE-764326-05"] <- 1951

SHARE_w3$year_of_birth <- car::recode(SHARE_w3$yrbirth, "-2=997; -1=998;  NA=999")


#harmonise month of birth
SHARE_w3$month_of_birth <- car::recode(SHARE_w3$mobirth, "-2=997; -1=998;  NA=999")

#harmonise month of interview
SHARE_w3$month_of_interview <- car::recode(SHARE_w3$int_month_w3, "-2=997; -1=998;  c(13,NA)=999")

#harmonise year of interview
SHARE_w3$year_of_interview <- car::recode(SHARE_w3$int_year_w3, "NA=2009")



SHARE_ds_w3$age <- comp_age(y_birth=SHARE_w3$year_of_birth , m_birth=SHARE_w3$month_of_birth ,
         y_int=SHARE_w3$year_of_interview , m_int=SHARE_w3$month_of_interview , miss=miss_values_vector)




```

#### Statistics of the new harmonised variable
```{r descript3, echo=F}
vari <- SHARE_ds_w3$age
kable(Continuous_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SHARE_ds_w3, aes(age)) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% miss_values_vector)) + xlab("Age") + ylab("Frequency")
```

#### Validation
```{r crosstabulation3, echo=F}
```








### Wave 4 

#### Study-specific variable description




| **Name** |`mobirth`| `yrbirth`| `int_month_w4` | `int_year_w4` |
|-|-|-|-|-|
| **Label** | `month of birth` | `year of birth`| `interview month` | `interview year` |
| **Table name** | `sharew4_rel1-1-1_cv_r_without_noint` | `idem` | `idem` | `idem` |
| **Categories** | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |  | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |  |
| **Missings** | `-2 = refusal`<br/>`-1 = don't know` | `-2 = refusal`<br/>`-1 = don't know`<br/>`NA` | `-2 = refusal`<br/>`-1 = don't know` | `-3 = implausible/ suspected wrong` |
| **Description** |  |  |  |  |

```{r assign4, echo=F}
opal.assign.table.tibble(o, 'SHARE_w4','SHARE.sharew4_rel1-1-1_cv_r_without_noint',
                         variables=list('mobirth', 'yrbirth', 'int_month_w4', 'int_year_w4'), missings = TRUE)
SHARE_w4 <- opal.execute(o,'SHARE_w4')
```

```{r local4, echo=F}
vari <- SHARE_w4$yrbirth
vbl_miss <- c(-2,-1)
kable(Continuous_summary(var = vari, missing_values = c(-2,-1))[3], caption = "Year of birth")
kable(Continuous_summary(vari, missing_values = c(-2,-1))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-2,-1))$summary, caption = "Summary")
ggplot(SHARE_w4, aes(yrbirth)) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% vbl_miss)) + xlab("Year of birth") + ylab("Frequency")

vari <- SHARE_w4$mobirth
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "Month of birth") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w4, aes(x=factor(mobirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")

vari <- SHARE_w4$int_year_w4
vbl_miss <- c(-2,-1)
kable(Continuous_summary(vari, missing_values = vbl_miss)[3], caption = "Year of interview") 
pander(Continuous_summary(vari, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SHARE_w4, aes(x=factor(int_year_w4))) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% vbl_miss)) + xlab("Year of interview") + ylab("Frequency")

vari <- SHARE_w4$int_month_w4
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "Month of interview") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w4, aes(x=factor(int_month_w4))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of interview") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variables it has to be computed by applying the following formula: 

* `date of interview - date of birth`

**R script:**

```{r harmo4}
SHARE_ds_w4 <- tibble(id=SHARE_w4$id)


#harmonise year of birth

#Some missing yrbirth data is available from previous waves
SHARE_w4 <- left_join(SHARE_w4,SHARE_w1 %>% select(id,yrbirth) %>% filter(!yrbirth %in% c(-1,-2)),by = "id", suffix = c("","_w1"))
SHARE_w4$yrbirth[which(is.na(SHARE_w4$yrbirth) | (SHARE_w4$yrbirth %in% c(-1,-2) & !is.na(SHARE_w4$yrbirth_w1)))] <- SHARE_w4$yrbirth_w1[which(is.na(SHARE_w4$yrbirth) | (SHARE_w4$yrbirth %in% c(-1,-2) & !is.na(SHARE_w4$yrbirth_w1)))]
SHARE_w4 <- left_join(SHARE_w4,SHARE_w2 %>% select(id,yrbirth) %>% filter(!yrbirth %in% c(-1,-2)),by = "id", suffix = c("","_w2"))
SHARE_w4$yrbirth[which(is.na(SHARE_w4$yrbirth) | (SHARE_w4$yrbirth %in% c(-1,-2) & !is.na(SHARE_w4$yrbirth_w2)))] <- SHARE_w4$yrbirth_w2[which(is.na(SHARE_w4$yrbirth) | (SHARE_w4$yrbirth %in% c(-1,-2) & !is.na(SHARE_w4$yrbirth_w2)))]
SHARE_w4 <- left_join(SHARE_w4,SHARE_w3 %>% select(id,yrbirth) %>% filter(!yrbirth %in% c(-1,-2)),by = "id", suffix = c("","_w3"))
SHARE_w4$yrbirth[which(is.na(SHARE_w4$yrbirth) | (SHARE_w4$yrbirth %in% c(-1,-2) & !is.na(SHARE_w4$yrbirth_w3)))] <- SHARE_w4$yrbirth_w3[which(is.na(SHARE_w4$yrbirth) | (SHARE_w4$yrbirth %in% c(-1,-2) & !is.na(SHARE_w4$yrbirth_w3)))]
SHARE_w4 <- SHARE_w4 %>% select(-yrbirth_w1,-yrbirth_w2,-yrbirth_w3)
yrbirth_missing <- union(yrbirth_missing,(SHARE_w4 %>% filter(is.na(yrbirth) | yrbirth %in% c(-1,-2)))$id)
SHARE_miss <- full_join(SHARE_miss,SHARE_w4 %>% select(id,yrbirth) %>% filter(id %in% id_missings),by="id",suffix = c("_w3","_w4"))
# Wave 4 missing yrbirth data is non-missing in posterior waves tables for some id's (look at validation of yrbirth at the end). Here is imputed.
SHARE_w4$yrbirth[which(SHARE_w4$id=="AT-864946-01")] <- 1945
SHARE_w4$yrbirth[which(SHARE_w4$id=="AT-886095-01")] <- 1926
SHARE_w4$yrbirth[which(SHARE_w4$id=="CZ-373473-04")] <- 1950
SHARE_w4$yrbirth[which(SHARE_w4$id=="CZ-694720-01")] <- 1954
SHARE_w4$yrbirth[which(SHARE_w4$id=="F1-191516-01")] <- 1950
SHARE_w4$yrbirth[which(SHARE_w4$id=="F1-728117-01")] <- 1946

SHARE_w4$yrbirth <- car::recode(SHARE_w4$yrbirth, "-2=997; -1=998; 2011=999;  NA=999")


#harmonise month of birth
SHARE_w4$mobirth <- car::recode(SHARE_w4$mobirth, "-2=997; -1=998;  NA=999")

#harmonise month of interview
SHARE_w4$int_month_w4 <- car::recode(SHARE_w4$int_month_w4, "-2=997; -1=998;  c(13,NA)=999")

#harmonise year of interview
SHARE_w4$int_year_w4 <- car::recode(SHARE_w4$int_year_w4,  "-3=999; NA=2011")



SHARE_ds_w4$age <- comp_age(y_birth=SHARE_w4$yrbirth , m_birth=SHARE_w4$mobirth ,
         y_int=SHARE_w4$int_year_w4 , m_int=SHARE_w4$int_month_w4 , miss=miss_values_vector)



```

#### Statistics of the new harmonised variable
```{r descript4, echo=F}
vari <- SHARE_ds_w4$age
kable(Continuous_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SHARE_ds_w4, aes(age)) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% miss_values_vector)) + xlab("Age") + ylab("Frequency")
```

#### Validation
```{r crosstabulation4, echo=F}
```








### Wave 5 

#### Study-specific variable description



| **Name** |`mobirth`| `yrbirth`| `int_month_w5` | `int_year_w5` |
|-|-|-|-|-|
| **Label** | `month of birth` | `year of birth`| `interview month` | `interview year` |
| **Table name** | `sharew5_rel1-0-0_cv_r_without_noint` | `idem` | `idem` | `idem` |
| **Categories** | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |  | `1 = january`<br/>`2 = february`<br/>`3 = march`<br/>`4 = april`<br/>`5 = may`<br/>`6 = june`<br/>`7 = july`<br/>`8 = august`<br/>`9 = september`<br/>`10 = october`<br/>`11 = november`<br/>`12 = december`<br/>`13 = today` |  |
| **Missings** | `-2 = refusal`<br/>`-1 = don't know` | `-2 = refusal`<br/>`-1 = don't know`<br/>`NA` | `-2 = refusal`<br/>`-1 = don't know` | `-2 = refusal`<br/>`-1 = don't know` |
| **Description** |  |  |  |  |



```{r assign5, echo=F}
opal.assign.table.tibble(o, 'SHARE_w5','SHARE.sharew5_rel1-0-0_cv_r_without_noint',
                         variables=list('yrbirth', 'mobirth', 'int_year_w5', 'int_month_w5'), missings = TRUE)
SHARE_w5 <- opal.execute(o,'SHARE_w5')
```

```{r local5, echo=F}
vari <- SHARE_w5$yrbirth
vbl_miss <- c(-2,-1)
kable(Continuous_summary(var = vari, missing_values = c(-2,-1))[3], caption = "Year of birth")
kable(Continuous_summary(vari, missing_values = c(-2,-1))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(-2,-1))$summary, caption = "Summary")
ggplot(SHARE_w5, aes(yrbirth)) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% vbl_miss)) + xlab("Year of birth") + ylab("Frequency")

vari <- SHARE_w5$mobirth
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "Month of birth") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w5, aes(x=factor(mobirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")

vari <- SHARE_w5$int_year_w5
vbl_miss <- c(-2,-1)
kable(Continuous_summary(vari, missing_values = vbl_miss)[3], caption = "Year of interview") 
pander(Continuous_summary(vari, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(SHARE_w5, aes(x=factor(int_year_w5))) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% vbl_miss)) + xlab("Year of interview") + ylab("Frequency")

vari <- SHARE_w5$int_month_w5
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "Month of interview") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w5, aes(x=factor(int_month_w5))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of interview") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variables it has to be computed by applying the following formula: 

* `date of interview - date of birth`

**R script:**

```{r harmo5}
SHARE_ds_w5 <- tibble(id=SHARE_w5$id)


#harmonise year of birth

#Some missing yrbirth data is available from previous waves
SHARE_w5 <- left_join(SHARE_w5,SHARE_w1 %>% select(id,yrbirth) %>% filter(!yrbirth %in% c(-1,-2)),by = "id", suffix = c("","_w1"))
SHARE_w5$yrbirth[which(is.na(SHARE_w5$yrbirth) | (SHARE_w5$yrbirth %in% c(-1,-2) & !is.na(SHARE_w5$yrbirth_w1)))] <- SHARE_w5$yrbirth_w1[which(is.na(SHARE_w5$yrbirth) | (SHARE_w5$yrbirth %in% c(-1,-2) & !is.na(SHARE_w5$yrbirth_w1)))]
SHARE_w5 <- left_join(SHARE_w5,SHARE_w2 %>% select(id,yrbirth) %>% filter(!yrbirth %in% c(-1,-2)),by = "id", suffix = c("","_w2"))
SHARE_w5$yrbirth[which(is.na(SHARE_w5$yrbirth) | (SHARE_w5$yrbirth %in% c(-1,-2) & !is.na(SHARE_w5$yrbirth_w2)))] <- SHARE_w5$yrbirth_w2[which(is.na(SHARE_w5$yrbirth) | (SHARE_w5$yrbirth %in% c(-1,-2) & !is.na(SHARE_w5$yrbirth_w2)))]
SHARE_w5 <- left_join(SHARE_w5,SHARE_w3 %>% select(id,yrbirth) %>% filter(!yrbirth %in% c(-1,-2)),by = "id", suffix = c("","_w3"))
SHARE_w5$yrbirth[which(is.na(SHARE_w5$yrbirth) | (SHARE_w5$yrbirth %in% c(-1,-2) & !is.na(SHARE_w5$yrbirth_w3)))] <- SHARE_w5$yrbirth_w3[which(is.na(SHARE_w5$yrbirth) | (SHARE_w5$yrbirth %in% c(-1,-2) & !is.na(SHARE_w5$yrbirth_w3)))]
SHARE_w5 <- left_join(SHARE_w5,SHARE_w4 %>% select(id,yrbirth) %>% filter(!yrbirth %in% c(-1,-2)),by = "id", suffix = c("","_w4"))
SHARE_w5$yrbirth[which(is.na(SHARE_w5$yrbirth) | (SHARE_w5$yrbirth %in% c(-1,-2) & !is.na(SHARE_w5$yrbirth_w4)))] <- SHARE_w5$yrbirth_w4[which(is.na(SHARE_w5$yrbirth) | (SHARE_w5$yrbirth %in% c(-1,-2) & !is.na(SHARE_w5$yrbirth_w4)))]
SHARE_w5 <- SHARE_w5 %>% select(-yrbirth_w1,-yrbirth_w2,-yrbirth_w3,-yrbirth_w4)
yrbirth_missing <- union(yrbirth_missing,(SHARE_w5 %>% filter(is.na(yrbirth) | yrbirth %in% c(-1,-2)))$id)
SHARE_miss <- full_join(SHARE_miss,SHARE_w5 %>% select(id,yrbirth) %>% filter(id %in% id_missings),by="id")

SHARE_w5$year_of_birth <- car::recode(SHARE_w5$yrbirth, "-2=997; -1=998; 2013=999;  NA=999")


#harmonise month of birth
SHARE_w5$month_of_birth <- car::recode(SHARE_w5$mobirth, "-2=997; -1=998;  NA=999")

#harmonise month of interview
SHARE_w5$month_of_interview <- car::recode(SHARE_w5$int_month_w5, "-2=997; -1=998;  13=999; NA=999")

#harmonise year of interview
SHARE_w5$year_of_interview <- car::recode(SHARE_w5$int_year_w5, "-2=997; -1=998; NA=999")



SHARE_ds_w5$age <- comp_age(y_birth=SHARE_w5$year_of_birth , m_birth=SHARE_w5$month_of_birth ,
         y_int=SHARE_w5$year_of_interview , m_int=SHARE_w5$month_of_interview , miss=miss_values_vector)




```

#### Statistics of the new harmonised variable
```{r descript5, echo=F}
vari <- SHARE_ds_w5$age
kable(Continuous_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(SHARE_ds_w5, aes(age)) + geom_bar(stat="count", width=0.4, fill="steelblue", data = .%>% filter(!vari %in% miss_values_vector)) + xlab("Age") + ylab("Frequency")
```

#### Validation
```{r crosstabulation5, echo=F}
```



### Validation of `yrbirth` across waves

```{r, echo=FALSE}

yrbirth_missing <- unique(yrbirth_missing)
cat("Validation of id's with previous missing yrbirth: ",identical(yrbirth_missing,id_missings))
kable(SHARE_miss %>% mutate(TNA=rowSums(is.na(.) | . == -1 | . == -2)) %>% filter(TNA!=5) %>% select(-TNA),caption = "Id's with yrbirth data but previous missing values")

```



## Summary of descriptive statistics of the harmonised variable accross populations and waves

```{r helpfunctions, echo=F}

source("../../helpfunctions.r")
rm(labelling)
labelling <- labelling_c
rm(summaries)
summaries <- summaries_c

```





```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names


share.cw <- list(w1 = c("w1","W1"), w2 = c("w2","W2"), w3 = c("w3","W3"), w4 = c("w4","W4"), w5 = c("w5","W5"))

# Consider only harmonised waves
l.hds <- lapply(share.cw, function(wname) if(exists(paste0("SHARE_ds_",wname[1]))){wname = list(get(paste0("SHARE_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing summaries
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```




# Quality estimation

Take into account the quality estimation of `ybirth` (harmonisation of `yrbirth`). There are years of birth not consistent across waves.



```{r, echo=FALSE}

share_age <- Reduce(full_join, lapply(l.hds, \(.x) rename_with(.x[[1]], .fn = \(.yn) paste(.yn, .x[[2]], sep = "_"), .cols = `age`)))
share_age |> 
  pivot_longer(cols = -id) |> 
  arrange(id,name) |> 
  filter(!is.na(value)) |> 
  group_by(id) |> 
  mutate(odd = any(value<=lag(value, default = -Inf))) |> 
  ungroup() |> 
  filter(odd) |> 
  mutate(name = sub("age_", "", name)) |> 
  pivot_wider(id_cols = "id", names_from = "name") |> 
  kable(caption = "Imposible age trajectories")
```



```{r closeRsession, echo=FALSE}
opal.logout(o)
```
