==> SHARE/1_Sociodemo_econom/3_sex.Rmd <==
---
title: "Sex"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/sexDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/sexDS.R')
```

<!-- ########################################################## --> 

# Data process



## SHARE 

### Wave 1 

#### Study-specific variable description

| **Name** | `gender` |
|----------|------------|
| **Label** | `T2000: gender, 1.7.2000` |
| **Table name** | `sharew1_rel2-6-0_dn` |
| **Categories** | `1 = male; 2 = female` |
| **Description** |  |


```{r assign1, echo=FALSE}
opal.assign.table.tibble(o, 'SHARE_w1_sex','SHARE.sharew1_rel2-6-0_dn',
                         variables=list('gender'), missings = TRUE)
SHARE_w1_sex <- opal.execute(o,'SHARE_w1_sex')
```

```{r local1, echo=FALSE}
vari <- SHARE_w1_sex$gender
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_w1_sex, aes(x=factor(gender))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("sex") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep "male" into 1`
* `2 "female" into 0`

**R script:**

```{r harmo1, echo=TRUE}
SHARE_ds_w1 <- tibble(id=SHARE_w1_sex$id)
SHARE_ds_w1$sex <- car::recode(SHARE_w1_sex$gender, "2=0; NA=999")
SHARE_ds_w1$sex <- labelled(SHARE_ds_w1$sex, labels = c("Female"=0, "Male"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript1, echo=FALSE}
vari <- SHARE_ds_w1$sex
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w1, aes(x=factor(sex))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("sex") + ylab("Frequency")
```

#### Validation
```{r crosstabulation1, echo=FALSE}
AfterH <- table(SHARE_ds_w1$sex)
BeforeH <- table(SHARE_w1_sex$gender)[c(2,1)]
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("2->0","1")
kable(C)
```

 
 



### Wave 2 

#### Study-specific variable description


| **Name** |`dn042_`|
|----------|---|
| **Label** | `male or female`|
| **Table name** |`sharew2_rel2-6-0_dn `| 
| **Categories** | `1 = male; 2 = female` |
| **Description** |  |


```{r assign2, echo=FALSE}

opal.assign.table.tibble(o, 'SHARE_w2_sex','SHARE.sharew2_rel2-6-0_dn',variables=list('dn042_'), missings = TRUE)
#opal.assign.table.tibble(o, 'SHARE_w1_sex','SHARE.sharew1_rel2-6-0_dn',variables=list('gender'), missings = TRUE)
# opal.assign.table.tibble(o, 'SHARE_w2_id','SHARE.sharew2_rel2-6-0_cv_r_without_noint',variables=list('hhid'), missings = TRUE)

SHARE_w2_sex <- opal.execute(o,'SHARE_w2_sex')
#SHARE_w1_sex <- opal.execute(o,'SHARE_w1_sex')
#SHARE_w2_id <- opal.execute(o,'SHARE_w2_id')
```

```{r local2, echo=FALSE}
# Add_indiv <- function(old, new){
#   
#   new_2 <- matrix(rep(NA, dim(new)[1]*dim(new)[2] ), nrow = dim(new)[1])
#   oldm <- as.matrix(old)
#   
#   for(i in 1:length(new$id)){
#     if(sum(new$id[i]==old$id)>=1){
#       new_2[i,] <- oldm[which(new$id[i]==old$id),]
#     } else {
#       new_2[i,] <- c(new$id[i], NA)
#     }
#   }
#   
#   new_22 <- data.frame(id=new_2[,1], gender=as.numeric(new_2[,2]))
#   new_22
# }
# 
# SHARE_w2_sex <- Add_indiv(old = SHARE_w1_sex, new = SHARE_w2_id)


vari <- SHARE_w2_sex$dn042_
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_w2_sex, aes(x=factor(dn042_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("sex") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep "male" into 1`
* `2 "female" into 0`

**R script:**

```{r harmo2, echo=TRUE}
SHARE_ds_w2 <- tibble(id=SHARE_w2_sex$id)
SHARE_ds_w2$sex <- car::recode(SHARE_w2_sex$dn042_, "2=0; NA=999")
SHARE_ds_w2$sex <- labelled(SHARE_ds_w2$sex, labels = c("Female"=0, "Male"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript2, echo=FALSE}
vari <- SHARE_ds_w2$sex
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w2, aes(x=factor(sex))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("sex") + ylab("Frequency")
```

#### Validation
```{r crosstabulation2, echo=FALSE}
AfterH <- table(SHARE_ds_w2$sex)
BeforeH <- c(table(SHARE_w2_sex$dn042_)[c(2,1)])
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("2->0","1")
kable(C)
```

 
 




### Wave 3 

#### Study-specific variable description


| **Name** | `gender` |`sl_st011_`|
|----------|----------|--|
| **Label** | `male or female` |`gender of respondent`|
| **Table name** | `sharew3_rel1_cv_r_without_noint` |`sharew3_rel1_st`|
| **Categories** | `1 = male; 2 = female` |`1 = male; 2 = female` |
| **Missings** |`NA`|`NA`|
| **Description** |  |  |


```{r assign3, echo=FALSE}

opal.assign.table.tibble(o, 'SHARE_w3_sex1','SHARE.sharew3_rel1_cv_r_without_noint',
                         variables=list('gender'), missings = TRUE)
SHARE_w3_sex1 <- opal.execute(o,'SHARE_w3_sex1')

opal.assign.table.tibble(o, 'SHARE_w3_sex2','SHARE.sharew3_rel1_st',
                         variables=list('sl_st011_'), missings = TRUE)
SHARE_w3_sex2 <- opal.execute(o,'SHARE_w3_sex2')

# opal.assign.table.tibble(o, 'SHARE_w1_sex','SHARE.sharew1_rel2-6-0_dn',variables=list('gender'), missings = TRUE)
# opal.assign.table.tibble(o, 'SHARE_w3_id','SHARE.sharew3_rel1_cv_r_without_noint',variables=list('hhid3'), missings = TRUE)
# SHARE_w1_sex <- opal.execute(o,'SHARE_w1_sex')
# SHARE_w3_id <- opal.execute(o,'SHARE_w3_id')
```

```{r local3, echo=FALSE}
# Add_indiv <- function(old, new){
#   
#   new_2 <- matrix(rep(NA, dim(new)[1]*dim(new)[2] ), nrow = dim(new)[1])
#   oldm <- as.matrix(old)
#   
#   for(i in 1:length(new$id)){
#     if(sum(new$id[i]==old$id)>=1){
#       new_2[i,] <- oldm[which(new$id[i]==old$id),]
#     } else {
#       new_2[i,] <- c(new$id[i], NA)
#     }
#   }
#   
#   new_22 <- data.frame(id=new_2[,1], gender=as.numeric(new_2[,2]))
#   new_22
# }
# 
# SHARE_w3_sex <- Add_indiv(old = SHARE_w1_sex, new = SHARE_w3_id)

vari <- SHARE_w3_sex1$gender
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_w3_sex1, aes(x=factor(gender))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("sex") + ylab("Frequency")

vari <- SHARE_w3_sex2$sl_st011_
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_w3_sex2, aes(x=factor(sl_st011_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("sex") + ylab("Frequency")


SHARE_w3_sex <- full_join(SHARE_w3_sex1,SHARE_w3_sex2,by="id",all = T)

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep "male" into 1`
* `2 "female" into 0`

Both study-specific variables are compatible in the sense that if none is missing they have the same value. Indeed, we check the number of differences: `r sum(SHARE_w3_sex$gender!=SHARE_w3_sex$sl_st011_,na.rm=T)`

**R script:**

```{r harmo3, echo=TRUE}
SHARE_ds_w3 <- tibble(id=SHARE_w3_sex$id)
SHARE_ds_w3$sex <- car::recode(SHARE_w3_sex$gender, "2=0; NA=999")
SHARE_ds_w3$sex[SHARE_ds_w3$sex == 999] <- car::recode(SHARE_w3_sex$sl_st011_[SHARE_ds_w3$sex == 999], "2=0; NA=999")
SHARE_ds_w3$sex <- labelled(SHARE_ds_w3$sex, labels = c("Female"=0, "Male"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript3, echo=FALSE}
vari <- SHARE_ds_w3$sex
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w3, aes(x=factor(sex))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("sex") + ylab("Frequency")
```

#### Validation
```{r crosstabulation3, echo=FALSE}
AfterH <- table(SHARE_ds_w3$sex)
BeforeH <- c(sum(SHARE_w3_sex$gender==2 | SHARE_w3_sex$sl_st011_==2,na.rm=T), sum(SHARE_w3_sex$gender==1 | SHARE_w3_sex$sl_st011_==1,na.rm=T), sum(is.na(SHARE_w3_sex$gender) & is.na(SHARE_w3_sex$sl_st011_)))
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("2->0","1", "NA->999")
kable(C)
```

 
 




### Wave 4 

#### Study-specific variable description



| **Name** | `gender` |`dn042_`|
|----------|----------|--|
| **Label** | `male or female` |`male or female`|
| **Table name** | `sharew4_rel1-1-1_cv_r_without_noint` |`sharew4_rel1-1-1_dn`|
| **Categories** | `1 = male; 2 = female` |`1 = male; 2 = female` |
| **Description** |  |  |



```{r assign4, echo=FALSE}
opal.assign.table.tibble(o, 'SHARE_w4_sex1','SHARE.sharew4_rel1-1-1_cv_r_without_noint',
                         variables=list('gender'), missings = TRUE)
SHARE_w4_sex1 <- opal.execute(o,'SHARE_w4_sex1')
opal.assign.table.tibble(o, 'SHARE_w4_sex2','SHARE.sharew4_rel1-1-1_dn',
                         variables=list('dn042_'), missings = TRUE)
SHARE_w4_sex2 <- opal.execute(o,'SHARE_w4_sex2')


#opal.assign.table.tibble(o, 'SHARE_w1_sex','SHARE.sharew1_rel2-6-0_dn', variables=list('gender'), missings = TRUE)
#opal.assign.table.tibble(o, 'SHARE_w4_id','SHARE.sharew4_rel1-1-1_cv_r_without_noint',variables=list('hhid4'), missings = TRUE)
#SHARE_w1_sex <- opal.execute(o,'SHARE_w1_sex')
#SHARE_w4_id <- opal.execute(o,'SHARE_w4_id')
```

```{r local4, echo=FALSE}
# Add_indiv <- function(old, new){
#   
#   new_2 <- matrix(rep(NA, dim(new)[1]*dim(new)[2] ), nrow = dim(new)[1])
#   oldm <- as.matrix(old)
#   
#   for(i in 1:length(new$id)){
#     if(sum(new$id[i]==old$id)>=1){
#       new_2[i,] <- oldm[which(new$id[i]==old$id),]
#     } else {
#       new_2[i,] <- c(new$id[i], NA)
#     }
#   }
#   
#   new_22 <- data.frame(id=new_2[,1], gender=as.numeric(new_2[,2]))
#   new_22
# }
# 
# SHARE_w4_sex <- Add_indiv(old = SHARE_w1_sex, new = SHARE_w4_id)

vari <- SHARE_w4_sex1$gender
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_w4_sex1, aes(x=factor(gender))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("sex") + ylab("Frequency")

vari <- SHARE_w4_sex2$dn042_
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_w4_sex2, aes(x=factor(dn042_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("sex") + ylab("Frequency")


SHARE_w4_sex <- full_join(SHARE_w4_sex1,SHARE_w4_sex2,by="id",all = T)
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep "male" into 1`
* `2 "female" into 0`

Both study-specific variables are compatible in the sense that if none is missing they have the same value. Indeed, we check the number of differences: `r sum(SHARE_w4_sex$gender!=SHARE_w4_sex$dn042_,na.rm=T)`

**R script:**

```{r harmo4, echo=TRUE}
SHARE_ds_w4 <- tibble(id=SHARE_w4_sex$id)
SHARE_ds_w4$sex <- car::recode(SHARE_w4_sex$gender, "2=0; NA=999")
SHARE_ds_w4$sex[SHARE_ds_w4$sex == 999] <- car::recode(SHARE_w4_sex$dn042_[SHARE_ds_w4$sex == 999], "2=0; NA=999")
SHARE_ds_w4$sex <- labelled(SHARE_ds_w4$sex, labels = c("Female"=0, "Male"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript4, echo=FALSE}
vari <- SHARE_ds_w4$sex
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w4, aes(x=factor(sex))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("sex") + ylab("Frequency")
```

#### Validation
```{r crosstabulation4, echo=FALSE}
AfterH <- table(SHARE_ds_w4$sex)
BeforeH <- c(sum(SHARE_w4_sex$gender==2 | SHARE_w4_sex$dn042_==2,na.rm=T),sum(SHARE_w4_sex$gender==1 | SHARE_w4_sex$dn042_==1,na.rm=T))
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("2->0","1")
kable(C)
```

 
 






### Wave 5 

#### Study-specific variable description



| **Name** | `gender` |`gender_do`|
|----------|-----|----|---|
| **Label** | `Male or female` |`Gender`|
| **Table name** | `sharew5_rel1-0-0_cv_r_without_noint` |`sharew5_rel1-0-0_do`|
| **Categories** | `1 = Male; 2 = Female` |`1 = Male; 2 = Female` |
| **Description** |  |  |


```{r assign5, echo=FALSE}
opal.assign.table.tibble(o, 'SHARE_w5_sex1','SHARE.sharew5_rel1-0-0_cv_r_without_noint',variables=list('gender'), missings = TRUE)
opal.assign.table.tibble(o, 'SHARE_w5_sex2','SHARE.sharew5_rel1-0-0_do',variables=list('gender_do'), missings = TRUE)

SHARE_w5_sex1 <- opal.execute(o,'SHARE_w5_sex1')
SHARE_w5_sex2 <- opal.execute(o,'SHARE_w5_sex2')


# opal.assign.table.tibble(o, 'SHARE_w1_sex','SHARE.sharew1_rel2-6-0_dn',variables=list('gender'), missings = TRUE)
# opal.assign.table.tibble(o, 'SHARE_w5_id','SHARE.sharew5_rel1-0-0_cv_r_without_noint', variables=list('hhid5'), missings = TRUE)
# SHARE_w1_sex <- opal.execute(o,'SHARE_w1_sex')
# SHARE_w5_id <- opal.execute(o,'SHARE_w5_id')
```

```{r local5, echo=FALSE}
# Add_indiv <- function(old, new){
#   
#   new_2 <- matrix(rep(NA, dim(new)[1]*dim(new)[2] ), nrow = dim(new)[1])
#   oldm <- as.matrix(old)
#   
#   for(i in 1:length(new$id)){
#     if(sum(new$id[i]==old$id)>=1){
#       new_2[i,] <- oldm[which(new$id[i]==old$id),]
#     } else {
#       new_2[i,] <- c(new$id[i], NA)
#     }
#   }
#   
#   new_22 <- data.frame(id=new_2[,1], gender=as.numeric(new_2[,2]))
#   new_22
# }
# 
# SHARE_w5_sex <- Add_indiv(old = SHARE_w1_sex, new = SHARE_w5_id)

vari <- SHARE_w5_sex1$gender
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_w5_sex1, aes(x=factor(gender))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("sex") + ylab("Frequency")

vari <- SHARE_w5_sex2$gender_do
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_w5_sex2, aes(x=factor(gender_do))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("sex") + ylab("Frequency")


SHARE_w5_sex <- full_join(SHARE_w5_sex1,SHARE_w5_sex2,by="id",all = T)


```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep "male" into 1`
* `2 "female" into 0`

Both study-specific variables are compatible in the sense that if none is missing they have the same value. Indeed, we check the number of differences: `r sum(SHARE_w5_sex$gender!=SHARE_w5_sex$gender_do,na.rm=T)`


**R script:**

```{r harmo5, echo=TRUE}
SHARE_ds_w5 <- tibble(id=SHARE_w5_sex$id)
SHARE_ds_w5$sex <- car::recode(SHARE_w5_sex$gender, "2=0; NA=999")
SHARE_ds_w5$sex[SHARE_ds_w5$sex == 999] <- car::recode(SHARE_w5_sex$gender_do[SHARE_ds_w5$sex == 999], "2=0; NA=999")
SHARE_ds_w5$sex <- labelled(SHARE_ds_w5$sex, labels = c("Female"=0, "Male"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript5, echo=FALSE}
vari <- SHARE_ds_w5$sex
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w5, aes(x=factor(sex))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("sex") + ylab("Frequency")
```

#### Validation
```{r crosstabulation5, echo=FALSE}
AfterH <- table(SHARE_ds_w5$sex)
BeforeH <- c(table(SHARE_w5_sex$gender)[c(2,1)])
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("2->0","1")
kable(C)
```

 




## Summary of descriptive statistics of the harmonised variable accross populations and waves

Two tables are generated: 

1. Percentages of categories in each harmonised variable.

2. Frequencies and percentages of individuals with different trajectories.



```{r helpfunctions, echo=F}

source("../../helpfunctions.r")

```




```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names



share.cw <- list(w1 = c("w1","W1"), w2 = c("w2","W2"), w3 = c("w3","W3"), w4 = c("w4","W4"), w5 = c("w5","W5"))

# Consider only harmonised waves
l.hds <- lapply(share.cw, function(wname) if(exists(paste0("SHARE_ds_",wname[1]))){wname = list(get(paste0("SHARE_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing table of categories
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Printing table of trajectories
  f.dbb <- trajectories(lmlist[[2]],vbl_name=hd_vbl)
  print(knitr::kable(f.dbb))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```






# Quality estimation

```{r, echo=F}

SHARE <- full_join(SHARE_ds_w1[which(!SHARE_ds_w1$sex %in% miss_values_vector),],SHARE_ds_w2[which(!SHARE_ds_w2$sex %in% miss_values_vector),], by="id", suffix=c(".w1",".w2"))
SHARE <- full_join(SHARE,SHARE_ds_w3[which(!SHARE_ds_w3$sex %in% miss_values_vector),], by="id")
SHARE <- full_join(SHARE,SHARE_ds_w4[which(!SHARE_ds_w4$sex %in% miss_values_vector),], by="id", suffix=c(".w3",".w4"))
SHARE <- full_join(SHARE,SHARE_ds_w5[which(!SHARE_ds_w5$sex %in% miss_values_vector),], by="id")
SHARE$var <- apply(SHARE[,-1],1,function(x) var(x, na.rm = T))

SHstr <- paste("There are ",length(SHARE$id[which(SHARE$var!=0)])," individuals with sex changing across waves.")

kable(frq(SHARE$var==0), align = 'c', caption = SHstr)
kable(SHARE[which(SHARE$var>0 & SHARE$var!=0),], align = 'c', caption = 'The trajectories of the 11 individuals with sex data changing across waves.')

```

We do not anything thereon.



```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
