==> SHARE/1_Sociodemo_econom/4_marital_status.Rmd <==
---
title: "Marital Status"
output: 
  html_document:
    theme: united
    highlight: tango
    toc: true
    toc_depth: 6
    toc_float:
      collapsed: true
      smooth_scroll: true
    number_sections: true
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->



```{r setup, include=FALSE,  echo=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")

```

<!-- ########################################################## --> 

# Description of DataSchema variable 

* Short name: `marital_status`
* Variable label: `The current marital status of the participant`
* Domain: `Socio-demographic and economic characteristics`
* Value type: `categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
1        | single
2        | married-cohabitating
3        | divorced-separated
4        | widow

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing



```{r lab, echo=FALSE}
ds_label <- "Marital status"
ds_label_all <- "The current marital status of the participant"
```



<!-- ########################################################## --> 

# Data process

## SHARE

### Wave 1 

#### Study-specific variable description



| **Name** | `dn014_` |
|----------|------------|
| **Label** | `marital status` |
| **Table name** | `sharew1_rel2-6-0_dn` |
| **Categories** | `1 = married and living together with spouse; 2 = registered partnership; 3 = married, living separated from spouse; 4 = never married; 5 = divorced; 6 = widowed` |
| **Missings** | `-2 = refusal; -1 = don't know` |
| **Description** |  |



```{r assign1,  echo=FALSE}
opal.assign.table.tibble(o, 'SHARE_w1_ms','SHARE.sharew1_rel2-6-0_dn',
                         variables=list('dn014_'), missings = TRUE)
SHARE_w1_ms <- opal.execute(o,'SHARE_w1_ms')
```

```{r local1,  echo=FALSE}
vari <- SHARE_w1_ms$dn014_
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w1_ms, aes(x=factor(dn014_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Marital Status") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `recode 1 "married and living together with spouse" into 2`
* `keep "registered partnership" into 2`
* `recode 3 "married, living separated from spouse" into 2`
* `recode 4 "never married" into 1`
* `recode 5 "divorced" into 3`
* `recode 6 "widowed" into 4`
* `recode -1 "don't know" into 998`
* `recode -2 "refusal" into 997`
* `recode NA into 999`

**R script:**

```{r harmo1}
SHARE_ds_w1 <- tibble(id=SHARE_w1_ms$id)
SHARE_ds_w1$marital_status <- car::recode(SHARE_w1_ms$dn014_, "1=2; 4=1; 5=3; 6=4; 3=2; -1=998; -2=997; NA='999'")
SHARE_ds_w1$marital_status <- labelled(SHARE_ds_w1$marital_status, labels = c("single"=1, "married cohabitating"=2, "divorced separated"=3, "widow"=4, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript1,  echo=FALSE}
vari <- SHARE_ds_w1$marital_status
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w1, aes(x=factor(marital_status))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("marital_status") + ylab("Frequency")
```

#### Validation
```{r crosstabulation1,  echo=FALSE}
BeforeH <- table(SHARE_w1_ms$dn014_)
BeforeH <- c(BeforeH[6], sum(BeforeH[c(3,4,5)]), BeforeH[7], BeforeH[8], BeforeH[1], BeforeH[2],  sum(is.na(SHARE_w1_ms$dn014_)) )
AfterH <- table(SHARE_ds_w1$marital_status)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("4->1","(1,2,3)->2", "5->3", "6->4","-2->997", "-1->998", "NA->999")
kable(C)
```

 
 

 
```{r importOPAL1,  echo=FALSE}
label(SHARE_ds_w1$marital_status) <- ds_label_all
marital_status <- SHARE_ds_w1
save(marital_status, file = "../RData/w1/marital_status.RData")
#opal.assign.data(o, 'SHARE_ds_w1', SHARE_ds_w1)
#opal.symbol_import(o,'SHARE_ds_w1', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w1 <- SHARE_ds_w1
SHARE_m_ds_w1$marital_status <- car::recode(SHARE_m_ds_w1$marital_status, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w1', SHARE_m_ds_w1)
#opal.symbol_import(o,'SHARE_m_ds_w1', project='_Athlos_Harm_Dataset')
```






### Wave 2 

#### Study-specific variable description

 
| **Name** | `dn014_` |`dn044_`| `marital_status` |
|----------|--------|--|--|
| **Label** | `marital status` |`marital status changed`|`The current marital status of the participant` |
| **Table name** | `sharew2_rel2-6-0_dn` | `sharew2_rel2-6-0_dn` |`marital_status` |
| **Categories** | `1 = married and living together with spouse; 2 = registered partnership; 3 = married, living separated from spouse; 4 = never married; 5 = divorced; 6 = widowed` | `1 = yes, marital status has changed; 5 = no, marital status has not changed` |`1 = single; 2 = married-cohabitating; 3 = divorced-separated; 4 = widow` |
| **Missings** | `-2 = refusal; -1 = don't know` | `-2 = refusal; -1 = don't know` |`991 = CAPI/interviewer error; 992 = Impute; 993 = Disable to measure; 994 = Not attempt/not done; 995 = Does not answer; 996 = Not applicable; 997 = Refuse; 998 = Do not know; 999 = Missing` |
| **Description** |  |  |  |





```{r assign2,  echo=FALSE}
opal.assign.table.tibble(o, 'SHARE_w2_ms','SHARE.sharew2_rel2-6-0_dn', variables=list('dn014_','dn044_'), missings = TRUE)
SHARE_w2_ms <- opal.execute(o,'SHARE_w2_ms')
load("../RData/w1/marital_status.RData")
SHARE_ds_w1 <- marital_status
#opal.assign.table.tibble(o, 'SHARE_w1_ms','SHARE.sharew1_rel2-6-0_dn',variables=list('dn014_'), missings = TRUE)
#SHARE_w1_ms <- opal.execute(o,'SHARE_w1_ms')
```

```{r local2,  echo=FALSE}
vari <- SHARE_w2_ms$dn014_
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w2_ms, aes(x=factor(dn014_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Marital Status") + ylab("Frequency")

vari <- SHARE_w2_ms$dn044_
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w2_ms, aes(x=factor(dn044_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Marital Status") + ylab("Frequency")

# vari <- SHARE_w1_ms$dn014_
# kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = ds_label) 
# kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
# kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
# ggplot(SHARE_w1_ms, aes(x=factor(dn014_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Marital Status") + ylab("Frequency")
```





#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `recode 1 "married and living together with spouse" into 2`
* `keep "registered partnership" into 2`
* `recode 3 "married, living separated from spouse" into 2`
* `recode 4 "never married" into 1`
* `recode 5 "divorced" into 3`
* `recode 6 "widowed" into 4`
* `recode -1 "don't know" into 998`
* `recode -2 "refusal" into 997`
* `recode NA into 999`
* `if dn044_==5, marital_status into wave 1 data`


**R script:**


```{r harmo2}
SHARE_ds_w2 <- tibble(id=SHARE_w2_ms$id)
SHARE_ds_w2$marital_status <- car::recode(SHARE_w2_ms$dn014_, "1=2; 4=1;3=2; 5=3; 6=4; -1=998; -2=997; NA='999'")


assign_merge <- function(endtib,sttib,col_end_name="",col_st_name="",log_vec=c(T),col_end_index=NULL,col_st_index=NULL){
  newtib <- left_join(endtib,sttib,by="id")
  if(col_end_name!=""){
    col_end_index <- match(col_end_name,names(newtib))
  }
  if(col_st_name!=""){
    col_st_index <- match(col_st_name,names(newtib))
  }
  newtib[log_vec,col_end_index] <- newtib[log_vec,col_st_index]
  id_ind <- match("id",names(newtib))
  newtib <- newtib[,c(id_ind,col_end_index)]
  names(newtib)[2] <- "x"
  newtib$x <- car::recode(newtib$x,"NA=999")
  names(newtib)[2] <- names(endtib)[col_end_index]
  newtib
}


SHARE_ds_w2 <- assign_merge(SHARE_ds_w2,SHARE_ds_w1,log_vec = (SHARE_w2_ms$dn044_==5  & !is.na(SHARE_w2_ms$dn044_) & is.na(SHARE_w2_ms$dn014_)), col_end_index = 2,col_st_index = 3)
#SHARE_ds_w2$marital_status <- car::recode(SHARE_ds_w2$marital_status,"NA=999")


SHARE_ds_w2$marital_status <- labelled(SHARE_ds_w2$marital_status, labels = c("single"=1, "married cohabitating"=2, "divorced separated"=3, "widow"=4, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript2,  echo=FALSE}
vari <- SHARE_ds_w2$marital_status
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w2, aes(x=factor(marital_status))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("marital_status") + ylab("Frequency")
```

#### Validation
```{r crosstabulation2,  echo=FALSE}
BeforeH <- table(SHARE_w2_ms$dn014_)
BeforeHH <- table(SHARE_w1_ms$dn014_[SHARE_w1_ms$id %in% SHARE_w2_ms$id[SHARE_w2_ms$dn044_==5  & !is.na(SHARE_w2_ms$dn044_) & is.na(SHARE_w2_ms$dn014_)]])

BeforeV <- c(BeforeH[6]+BeforeHH[6], sum(BeforeH[c(3,4)],BeforeHH[c(3,4)]), BeforeH[5]+BeforeHH[5], sum(BeforeH[c(7, 8)],BeforeHH[c(7,8)]), BeforeH[1]+BeforeHH[1], BeforeH[2]+BeforeHH[2],  sum(is.na(SHARE_w2_ms$dn014_))-sum(BeforeHH) )
AfterH <- table(SHARE_ds_w2$marital_status)
C <- rbind(BeforeV,AfterH)
colnames(C) <- c("4->1","(1,2,3)->2", "5->3", "6->4","-2->997", "-1->998", "NA->999")
kable(C)
```

 
 

 
```{r importOPAL2,  echo=FALSE}
label(SHARE_ds_w2$marital_status) <- ds_label_all
marital_status <- SHARE_ds_w2
save(marital_status, file = "../RData/w2/marital_status.RData")
#opal.assign.data(o, 'SHARE_ds_w2', SHARE_ds_w2)
#opal.symbol_import(o,'SHARE_ds_w2', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w2 <- SHARE_ds_w2
SHARE_m_ds_w2$marital_status <- car::recode(SHARE_m_ds_w2$marital_status, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w2', SHARE_m_ds_w2)
#opal.symbol_import(o,'SHARE_m_ds_w2', project='_Athlos_Harm_Dataset')
```




### Wave 3 

There is no variable.






### Wave 4 

#### Study-specific variable description

 
| **Name** | `dn014_` |`dn044_`| `marital_status` |
|----------|--------|--|--|
| **Label** | `marital status` |`marital status changed`|`The current marital status of the participant` |
| **Table name** | `sharew4_rel1-1-1_dn` | `sharew4_rel1-1-1_dn` |`marital_status` |
| **Categories** | `1 = married and living together with spouse; 2 = registered partnership; 3 = married, living separated from spouse; 4 = never married; 5 = divorced; 6 = widowed` | `1 = yes, marital status has changed; 5 = no, marital status has not changed` |`1 = single; 2 = married-cohabitating; 3 = divorced-separated; 4 = widow` |
| **Missings** | `-2 = refusal; -1 = don't know` | `-2 = refusal; -1 = don't know` |`991 = CAPI/interviewer error; 992 = Impute; 993 = Disable to measure; 994 = Not attempt/not done; 995 = Does not answer; 996 = Not applicable; 997 = Refuse; 998 = Do not know; 999 = Missing` |
| **Description** |  |  |  |


```{r assign4,  echo=FALSE}
opal.assign.table.tibble(o, 'SHARE_w4_ms','SHARE.sharew4_rel1-1-1_dn', variables=list('dn014_','dn044_'), missings = TRUE)
SHARE_w4_ms <- opal.execute(o,'SHARE_w4_ms')
load("../RData/w2/marital_status.RData")
SHARE_ds_w2 <- marital_status
```



```{r local4,  echo=FALSE}
# Add_indiv <- function(old, new){
#   
#   new_2 <- matrix(rep(NA, dim(new)[1]*dim(new)[2] ), nrow = dim(new)[1])
#   oldm <- as.matrix(old)
#   
#   for(i in 1:length(new$id)){
#     if(sum(new$id[i]==old$id)>=1){
#       new_2[i,] <- oldm[which(new$id[i]==old$id),]
#     } else {
#       new_2[i,] <- c(new$id[i], NA)
#     }
#   }
#   
#   new_22 <- data.frame(id=new_2[,1], dn014_=as.numeric(new_2[,2]))
#   new_22
# }
# 
# SHARE_w4_ms <- Add_indiv(old = SHARE_w4_ms, new = SHARE_w4_id)

vari <- SHARE_w4_ms$dn014_
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w4_ms, aes(x=factor(dn014_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Marital Status") + ylab("Frequency")

vari <- SHARE_w4_ms$dn044_
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w4_ms, aes(x=factor(dn044_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Marital Status") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `recode 1 "married and living together with spouse" into 2`
* `keep "registered partnership" into 2`
* `recode 3 "married, living separated from spouse" into 2`
* `recode 4 "never married" into 1`
* `recode 5 "divorced" into 3`
* `recode 6 "widowed" into 4`
* `recode -1 "don't know" into 998`
* `recode -2 "refusal" into 997`
* `recode NA into 999`
* `if dn044_==5, marital_status into wave 2 data`

**R script:**

```{r harmo4,   echo=TRUE}
SHARE_ds_w4 <- tibble(id=SHARE_w4_ms$id)
SHARE_ds_w4$marital_status <- car::recode(SHARE_w4_ms$dn014_, "1=2; 4=1; 5=3; 3=2; 6=4; -1=998; -2=997; NA=999")

assign_merge <- function(endtib,sttib,col_end_name="",col_st_name="",log_vec=c(T),col_end_index=NULL,col_st_index=NULL){
  newtib <- left_join(endtib,sttib,by="id")
  if(col_end_name!=""){
    col_end_index <- match(col_end_name,names(newtib))
  }
  if(col_st_name!=""){
    col_st_index <- match(col_st_name,names(newtib))
  }
  newtib[log_vec,col_end_index] <- newtib[log_vec,col_st_index]
  id_ind <- match("id",names(newtib))
  newtib <- newtib[,c(id_ind,col_end_index)]
  names(newtib)[2] <- "x"
  newtib$x <- car::recode(newtib$x,"NA=999")
  names(newtib)[2] <- names(endtib)[col_end_index]
  newtib
}

SHARE_ds_w4 <- assign_merge(SHARE_ds_w4,SHARE_ds_w2,log_vec = (SHARE_w4_ms$dn044_==5  & !is.na(SHARE_w4_ms$dn044_) & is.na(SHARE_w4_ms$dn014_)), col_end_index = 2,col_st_index = 3)

#SHARE_ds_w4$marital_status <- car::recode(SHARE_ds_w4$marital_status,"NA=999")

SHARE_ds_w4$marital_status <- labelled(SHARE_ds_w4$marital_status, labels = c("single"=1, "married cohabitating"=2, "divorced separated"=3, "widow"=4, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript4,  echo=FALSE}
vari <- SHARE_ds_w4$marital_status
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w4, aes(x=factor(marital_status))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("marital_status") + ylab("Frequency")
```

#### Validation
```{r crosstabulation4,  echo=FALSE}
BeforeH <- table(SHARE_w4_ms$dn014_)

BeforeHH <- table(SHARE_ds_w2$marital_status[SHARE_ds_w2$id %in% SHARE_w4_ms$id[SHARE_w4_ms$dn044_==5  & !is.na(SHARE_w4_ms$dn044_) & is.na(SHARE_w4_ms$dn014_)]])


BeforeH <- c(BeforeH[6]+BeforeHH[1], sum(BeforeH[c(3,4)])+BeforeHH[2], BeforeH[5]+BeforeHH[3], sum(BeforeH[c(7, 8)],BeforeHH[4]), BeforeH[1]+BeforeHH[5], BeforeH[2]+BeforeHH[6],  sum(is.na(SHARE_w4_ms$dn014_))-sum(BeforeHH[1:6]) )
AfterH <- table(SHARE_ds_w4$marital_status)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("4->1","(1,2)->2", "3", "(5,6)->4","-2->997", "-1->998", "NA->999")
kable(C)
```

 
 

 
```{r importOPAL4,  echo=FALSE}
label(SHARE_ds_w4$marital_status) <- ds_label_all
marital_status <- SHARE_ds_w4
save(marital_status, file = "../RData/w4/marital_status.RData")
#opal.assign.data(o, 'SHARE_ds_w4', SHARE_ds_w4)
#opal.symbol_import(o,'SHARE_ds_w4', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w4 <- SHARE_ds_w4
SHARE_m_ds_w4$marital_status <- car::recode(SHARE_m_ds_w4$marital_status, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w4', SHARE_m_ds_w4)
#opal.symbol_import(o,'SHARE_m_ds_w4', project='_Athlos_Harm_Dataset')
```





### Wave 5 

#### Study-specific variable description

  
| **Name** | `dn014_` |`dn044_`| `marital_status` |
|----------|--------|--|--|
| **Label** | `Marital status` |`Marital status changed`|`The current marital status of the participant` |
| **Table name** | `sharew5_rel1-0-0_dn` | `sharew5_rel1-0-0_dn` |`marital_status` |
| **Categories** | `1 = married and living together with spouse; 2 = registered partnership; 3 = married, living separated from spouse; 4 = never married; 5 = divorced; 6 = widowed` | `1 = yes, marital status has changed; 5 = no, marital status has not changed` |`1 = single; 2 = married-cohabitating; 3 = divorced-separated; 4 = widow` |
| **Missings** | `-2 = refusal; -1 = don't know` | `-2 = refusal; -1 = don't know` |`991 = CAPI/interviewer error; 992 = Impute; 993 = Disable to measure; 994 = Not attempt/not done; 995 = Does not answer; 996 = Not applicable; 997 = Refuse; 998 = Do not know; 999 = Missing` |
| **Description** |  |  |  |





```{r assign5,  echo=FALSE}
opal.assign.table.tibble(o, 'SHARE_w5_ms','SHARE.sharew5_rel1-0-0_dn', variables=list('dn014_','dn044_'), missings = TRUE)
SHARE_w5_ms <- opal.execute(o,'SHARE_w5_ms')
load("../RData/w4/marital_status.RData")
SHARE_ds_w4 <- marital_status
```



```{r local5,  echo=FALSE}
# Add_indiv <- function(old, new){
#   
#   new_2 <- matrix(rep(NA, dim(new)[1]*dim(new)[2] ), nrow = dim(new)[1])
#   oldm <- as.matrix(old)
#   
#   for(i in 1:length(new$id)){
#     if(sum(new$id[i]==old$id)>=1){
#       new_2[i,] <- oldm[which(new$id[i]==old$id),]
#     } else {
#       new_2[i,] <- c(new$id[i], NA)
#     }
#   }
#   
#   new_22 <- data.frame(id=new_2[,1], dn014_=as.numeric(new_2[,2]))
#   new_22
# }

#SHARE_w5_ms <- Add_indiv(old = SHARE_w5_ms, new = SHARE_w5_id)

vari <- SHARE_w5_ms$dn014_
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w5_ms, aes(x=factor(dn014_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Marital Status") + ylab("Frequency")
vari <- SHARE_w5_ms$dn044_
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w5_ms, aes(x=factor(dn044_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Marital Status") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `recode 1 "married and living together with spouse" into 2`
* `keep "registered partnership" into 2`
* `recode 3 "married, living separated from spouse" into 2`
* `recode 4 "never married" into 1`
* `recode 5 "divorced" into 3`
* `recode 6 "widowed" into 4`
* `recode -1 "don't know" into 998`
* `recode -2 "refusal" into 997`
* `recode NA into 999`
* `if dn044_==5, marital_status into wave 4 data`

**R script:**

```{r harmo5,   echo=TRUE}
SHARE_ds_w5 <- tibble(id=SHARE_w5_ms$id)
SHARE_ds_w5$marital_status <- car::recode(SHARE_w5_ms$dn014_, "1=2; 4=1; 5=3; 3=2; 6=4; -1=998; -2=997; NA=999")

assign_merge <- function(endtib,sttib,col_end_name="",col_st_name="",log_vec=c(T),col_end_index=NULL,col_st_index=NULL){
  newtib <- left_join(endtib,sttib,by="id")
  if(col_end_name!=""){
    col_end_index <- match(col_end_name,names(newtib))
  }
  if(col_st_name!=""){
    col_st_index <- match(col_st_name,names(newtib))
  }
  newtib[log_vec,col_end_index] <- newtib[log_vec,col_st_index]
  id_ind <- match("id",names(newtib))
  newtib <- newtib[,c(id_ind,col_end_index)]
  names(newtib)[2] <- "x"
  newtib$x <- car::recode(newtib$x,"NA=999")
  names(newtib)[2] <- names(endtib)[col_end_index]
  newtib
}

SHARE_ds_w5 <- assign_merge(SHARE_ds_w5,SHARE_ds_w4,log_vec = (SHARE_w5_ms$dn044_==5  & !is.na(SHARE_w5_ms$dn044_) & is.na(SHARE_w5_ms$dn014_)), col_end_index = 2,col_st_index = 3)

#SHARE_ds_w5$marital_status <- car::recode(SHARE_ds_w5$marital_status,"NA=999")


SHARE_ds_w5$marital_status <- labelled(SHARE_ds_w5$marital_status, labels = c("single"=1, "married cohabitating"=2, "divorced separated"=3, "widow"=4, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript5,  echo=FALSE}
vari <- SHARE_ds_w5$marital_status
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w5, aes(x=factor(marital_status))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("marital_status") + ylab("Frequency")
```

#### Validation
```{r crosstabulation5,  echo=FALSE}
BeforeH <- table(SHARE_w5_ms$dn014_)

BeforeHH <- table(SHARE_ds_w4$marital_status[SHARE_ds_w4$id %in% SHARE_w5_ms$id[SHARE_w5_ms$dn044_==5  & !is.na(SHARE_w5_ms$dn044_) & is.na(SHARE_w5_ms$dn014_)]])


BeforeH <- c(BeforeH[6]+BeforeHH[1], sum(BeforeH[c(3,4)])+BeforeHH[2], BeforeH[5]+BeforeHH[3], sum(BeforeH[c(7, 8)],BeforeHH[4]), BeforeH[1]+BeforeHH[5], BeforeH[2]+BeforeHH[6],  sum(is.na(SHARE_w5_ms$dn014_))-sum(BeforeHH[1:6]) )

AfterH <- table(SHARE_ds_w5$marital_status)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("4->1","(1,2)->2", "3", "(5,6)->4","-2->997", "-1->998", "NA->999")
kable(C)
```

 
 

 
```{r importOPAL5,  echo=FALSE}
label(SHARE_ds_w5$marital_status) <- ds_label_all
marital_status <- SHARE_ds_w5
save(marital_status, file = "../RData/w5/marital_status.RData")
#opal.assign.data(o, 'SHARE_ds_w5', SHARE_ds_w5)
#opal.symbol_import(o,'SHARE_ds_w5', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w5 <- SHARE_ds_w5
SHARE_m_ds_w5$marital_status <- car::recode(SHARE_m_ds_w5$marital_status, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w5', SHARE_m_ds_w5)
#opal.symbol_import(o,'SHARE_m_ds_w5', project='_Athlos_Harm_Dataset')
```

## Summary of descriptive statistics of the harmonised variable accross populations and waves

Two tables are generated: 

1. Percentages of categories in each harmonised variable.

2. Frequencies and percentages of individuals with different trajectories. To avoid too large table, Only trajectories with percentages larger than 0.3 are shown.

**R script:**

"single"=1, "married cohabitating"=2, "divorced separated"=3, "widow"=4

```{r summ}
SHARE_ds_w1$marital_status <- labelled(SHARE_ds_w1$marital_status, labels = c(single=1,  married_cohabitating=2, divorced_separated=3, widow=4, Missing=999, Do_not_know=998, Refuse=997, Not_applicable=996, Does_not_answer=995,  Not_attempt_not_done=994, Disable_to_measure=993, Impute=992, CAPI_interviewer_error=991))
SHARE_ds_w2$marital_status <- labelled(SHARE_ds_w2$marital_status, labels = c(single=1,  married_cohabitating=2, divorced_separated=3, widow=4, Missing=999, Do_not_know=998, Refuse=997, Not_applicable=996, Does_not_answer=995,  Not_attempt_not_done=994, Disable_to_measure=993, Impute=992, CAPI_interviewer_error=991))
SHARE_ds_w4$marital_status <- labelled(SHARE_ds_w4$marital_status, labels = c(single=1,  married_cohabitating=2, divorced_separated=3, widow=4, Missing=999, Do_not_know=998, Refuse=997, Not_applicable=996, Does_not_answer=995,  Not_attempt_not_done=994, Disable_to_measure=993, Impute=992, CAPI_interviewer_error=991))
SHARE_ds_w5$marital_status <- labelled(SHARE_ds_w5$marital_status, labels = c(single=1,  married_cohabitating=2, divorced_separated=3, widow=4, Missing=999, Do_not_know=998, Refuse=997, Not_applicable=996, Does_not_answer=995,  Not_attempt_not_done=994, Disable_to_measure=993, Impute=992, CAPI_interviewer_error=991))

tt <-    frq(SHARE_ds_w1$marital_status)[[1]][,c(1:2,4)]
tt[4] <- frq(SHARE_ds_w2$marital_status)[[1]][,c(4)]
tt[5] <- frq(SHARE_ds_w4$marital_status)[[1]][,c(4)]
tt[6] <- frq(SHARE_ds_w5$marital_status)[[1]][,c(4)]
tt[13,] <- c("n", "sample size", 
             length(SHARE_ds_w1$marital_status), length(SHARE_ds_w2$marital_status), 
             length(SHARE_ds_w4$marital_status), length(SHARE_ds_w5$marital_status)
             )
names(tt)<-c("val", "label", "w1", "w2","w4", "w5")
kable(tt)

dbb <- merge(SHARE_m_ds_w1, SHARE_m_ds_w2, by="id", suffixes = c(".w1",".w2"), all=T)
dbb <- merge(dbb, SHARE_m_ds_w4, by="id", all=T)
dbb <- merge(dbb, SHARE_m_ds_w5, by="id", suffixes = c(".w4",".w5"), all=T)
names(dbb) <- c("id", "w1", "w2", "w4", "w5")
v.dbb <- paste(dbb[,2],dbb[,3],dbb[,4],dbb[,5],sep="")
f.dbb <- frq(v.dbb)[[1]][,c(1,2,4)]
kable(f.dbb[f.dbb[,3]>=0.3,])
```





# Quality estimation

'3=married, living separated from spouse' was considered as 2=married-cohabitaing label in the harmonized variable. 

[Comments on the quality of the new harmonised variable. This section can be done in a later stage, once all variables are harmonised.]

```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
