==> SHARE/1_Sociodemo_econom/5_education.Rmd <==
---
title: "Education"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/educationDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/educationDS.R')
```


<!-- ########################################################## --> 

# Data process

## SHARE 

### Wave 1 

#### Study-specific variable description

 
| **Name** | `isced_r` |
|----------|------------|
| **Label** | `ISCED-97 coding of education` |
| **Table name** | `sharew1_rel2-6-0_gv_isced` |
| **Categories** | `0 = none`<br/>`1 = isced-97 code 1`<br/>`2 = isced-97 code 2`<br/>`3 = isced-97 code 3`<br/>`4 = isced-97 code 4`<br/>`5 = isced-97 code 5`<br/>`6 = isced-97 code 6`<br/>`95 = still in school`<br/>`97 = other` |
| **Missings** | `-7 = not yet coded (temporary)`<br/>`-2 = refusal`<br/>`-1 = don't know` |
| **Description** |  |



```{r assign1,  echo=FALSE}
opal.assign.table.tibble(o, 'SHARE_w1_educ','SHARE.sharew1_rel2-6-0_gv_isced',
                         variables=list('isced_r'), missings = TRUE)
SHARE_w1_educ <- opal.execute(o,'SHARE_w1_educ')
```

```{r local1,  echo=FALSE}
vari <- SHARE_w1_educ$isced_r
kable(Categorical_summary(vari, missing_values = c(-7,-2,-1))[3], caption = "isced_r") 
kable(Categorical_summary(vari, missing_values = c(-7,-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-7,-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w1_educ, aes(x=factor(isced_r))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Level of education") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep "none" into 0`
* `keep "isced-97 code 1" into 1`
* `keep "isced-97 code 2" into 2`
* `recode 3 "isced-97 code 3" into 2`
* `recode 4 "isced-97 code 4" into 2`
* `recode 5 "isced-97 code 5" into 3`
* `recode 6 "isced-97 code 6" into 3`
* `recode 95 "still in school" into 996`
* `recode 97 "other" into 999`
* `recode -7 "not yet coded (temporary)" into 994`
* `recode -2 "refusal" into 997`
* `recode -1 "don't know" into 998`

**R script:**

```{r harmo1,   echo=TRUE}
SHARE_ds_w1 <- tibble(id=SHARE_w1_educ$id)
SHARE_ds_w1$education <- car::recode(SHARE_w1_educ$isced_r, "3=2; 4=2; 5=3; 6=3; 95=996; 97=999; -7=994;-2=997; -1=998;  NA=999")
SHARE_ds_w1$education <- labelled(SHARE_ds_w1$education, labels = c("less than primary education"=0, "primary education"=1, "secondary education"=2, "tertiary education"=3, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript1,  echo=FALSE}
vari <- SHARE_ds_w1$education
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w1, aes(x=factor(education))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Level of education") + ylab("Frequency")
```

#### Validation
```{r crosstabulation1,  echo=FALSE}
BeforeH <- table(SHARE_w1_educ$isced_r)
BeforeH <- c(BeforeH[3], BeforeH[4],sum(BeforeH[5:7]), sum(BeforeH[8:9]), BeforeH[10], BeforeH[1], BeforeH[2], sum(BeforeH[11],sum(is.na(SHARE_w1_educ$isced_r))))
AfterH <- table(SHARE_ds_w1$education)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("0", "1", "(2,3,4)->2", "(5,6)->3", "95->996", "-2->997", "-1->998", "(97,NA)->999")
kable(C)
```

 
 

 
```{r importOPAL1,  echo=FALSE}
label(SHARE_ds_w1$education) <- ds_label
education <- SHARE_ds_w1
save(education, file = "../RData/w1/education.RData")
#opal.assign.data(o, 'SHARE_ds_w1', SHARE_ds_w1)
#opal.symbol_import( o,'SHARE_ds_w1', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w1 <- SHARE_ds_w1
SHARE_m_ds_w1$education <- car::recode(SHARE_m_ds_w1$education, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w1', SHARE_m_ds_w1)
#opal.symbol_import( o,'SHARE_m_ds_w1', project='_Athlos_Harm_Dataset')
```





### Wave 2 

#### Study-specific variable description


 
| **Name** | `isced_r` |`mn101_`|
|----------|---------|---|
| **Label** | `ISCED-97 coding of education` |`questionnaire version`|
| **Table name** | `sharew2_rel2-6-0_gv_isced` |`sharew2_rel2-6-0_dn`|
| **Categories** | `0 = none`<br/>`1 = isced-97 code 1`<br/>`2 = isced-97 code 2`<br/>`3 = isced-97 code 3`<br/>`4 = isced-97 code 4`<br/>`5 = isced-97 code 5`<br/>`6 = isced-97 code 6`<br/>`95 = still in school`<br/>`97 = other` |`0 = baseline questionnaire version`<br/>`1 = longitudinal questionnaire version`|
| **Missings** | `-2 = refusal`<br/>`-1 = don't know` |`NA`|
| **Description** |  |  |





```{r assign2,  echo=FALSE}
opal.assign.table.tibble(o, 'SHARE_w2_educ','SHARE.sharew2_rel2-6-0_gv_isced',
                         variables=list('isced_r'), missings = TRUE)
SHARE_w2_educ <- opal.execute(o,'SHARE_w2_educ')

opal.assign.table.tibble(o, 'SHARE_w2','SHARE.sharew2_rel2-6-0_dn',
                         variables=list('mn101_'), missings = TRUE)
SHARE_w2 <- opal.execute(o,'SHARE_w2')

SHARE_w2_educ <- left_join(SHARE_w2_educ,SHARE_w2, by = "id")

```

```{r local2,  echo=FALSE}
vari <- SHARE_w2_educ$isced_r
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "isced_r") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w2_educ, aes(x=factor(isced_r))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Level of education") + ylab("Frequency")

vari <- SHARE_w2_educ$mn101_
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "mn101_") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
ggplot(SHARE_w2_educ, aes(x=factor(mn101_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vari)) + ylab("Frequency")

kable(table(SHARE_w2_educ$isced_r,SHARE_w2_educ$mn101_,useNA = "ifany"),caption = 'Variable mn101_ produces a skip')

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable `isced_r` it has to be recoded as follows:

* `keep "none" into 0`
* `keep "isced-97 code 1" into 1`
* `keep "isced-97 code 2" into 2`
* `recode 3 "isced-97 code 3" into 2`
* `recode 4 "isced-97 code 4" into 2`
* `recode 5 "isced-97 code 5" into 3`
* `recode 6 "isced-97 code 6" into 3`
* `recode 95 "still in school" into 996`
* `recode 97 "other" into 999`
* `recode -2 "refusal" into 997`
* `recode -1 "don't know" into 998`
* `missing values into wave 1 education if possible, otherwise, if mn101_==1, into 996`

**R script:**

```{r harmo2,   echo=TRUE}

SHARE_w2_educ <- left_join(SHARE_w2_educ,SHARE_ds_w1, by="id")
SHARE_ds_w2 <- tibble(id=SHARE_w2_educ$id)
SHARE_ds_w2$education <- car::recode(SHARE_w2_educ$isced_r, "3=2; 4=2; 5=3; 6=3; 95=996; 97=999; -2=997; -1=998;  NA=999")
SHARE_ds_w2$education[which(SHARE_w2_educ$mn101_==1)] <- 996
SHARE_ds_w2$education[which(is.na(SHARE_w2_educ$isced_r))] <- car::recode(SHARE_w2_educ$education[which(is.na(SHARE_w2_educ$isced_r))],"NA=999")
```

#### Statistical description of the new harmonised variable
```{r descript2,  echo=FALSE}
vari <- SHARE_ds_w2$education
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w2, aes(x=factor(education))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Level of education") + ylab("Frequency")
```

#### Validation
```{r crosstabulation2,  echo=FALSE}

```

 
 

 
```{r importOPAL2,  echo=FALSE}
label(SHARE_ds_w2$education) <- ds_label
education <- SHARE_ds_w2
save(education, file = "../RData/w2/education.RData")
#opal.assign.data(o, 'SHARE_ds_w2', SHARE_ds_w2)
#opal.symbol_import( o,'SHARE_ds_w2', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w2 <- SHARE_ds_w2
SHARE_m_ds_w2$education <- car::recode(SHARE_m_ds_w2$education, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w2', SHARE_m_ds_w2)
#opal.symbol_import( o,'SHARE_m_ds_w2', project='_Athlos_Harm_Dataset')
```




### Wave 3 

There is no variable.






### Wave 4 

#### Study-specific variable description

 
| **Name** | `isced_r` |`mn101_`|
|----------|---------|---|
| **Label** | `ISCED-97 coding of education` |`interview version`|
| **Table name** | `sharew4_rel1-1-1_gv_isced` |`sharew4_rel1-1-1_dn`|
| **Categories** | `0 = none`<br/>`1 = isced-97 code 1`<br/>`2 = isced-97 code 2`<br/>`3 = isced-97 code 3`<br/>`4 = isced-97 code 4`<br/>`5 = isced-97 code 5`<br/>`6 = isced-97 code 6`<br/>`95 = still in school`<br/>`97 = other` |`0 = baseline questionnaire`<br/>`1 = longitudinal questionnaire`|
| **Missings** | `-2 = refusal`<br/>`-1 = don't know` |`NA`|
| **Description** |  |  |



```{r assign4,  echo=FALSE}

opal.assign.table.tibble(o, 'SHARE_w4_educ','SHARE.sharew4_rel1-1-1_gv_isced', variables=list('isced_r'), missings = TRUE)
SHARE_w4_educ <- opal.execute(o,'SHARE_w4_educ')
opal.assign.table.tibble(o, 'SHARE_w4_id','SHARE.sharew4_rel1-1-1_cv_r_without_noint', variables=list('hhid4'), missings = TRUE)
SHARE_w4_id <- opal.execute(o,'SHARE_w4_id')
opal.assign.table.tibble(o, 'SHARE_w4','SHARE.sharew4_rel1-1-1_dn', variables=list('mn101_'), missings = TRUE)
SHARE_w4 <- opal.execute(o,'SHARE_w4')

SHARE_w4_educ <- left_join(SHARE_w4_educ,SHARE_w4, by = "id")
SHARE_w4_educ <- left_join(SHARE_w4_id,SHARE_w4_educ,by="id")
```

```{r local4,  echo=FALSE}

vari <- SHARE_w4_educ$isced_r
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "isced_r") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w4_educ, aes(x=factor(isced_r))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Level of education") + ylab("Frequency")


vari <- SHARE_w4_educ$mn101_
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "mn101_") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
ggplot(SHARE_w4_educ, aes(x=factor(mn101_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vari)) + ylab("Frequency")

kable(table(SHARE_w4_educ$isced_r,SHARE_w4_educ$mn101_,useNA = "ifany"),caption = 'Variable mn101_ produces a skip')

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep "none" into 0`
* `keep "isced-97 code 1" into 1`
* `keep "isced-97 code 2" into 2`
* `recode 3 "isced-97 code 3" into 2`
* `recode 4 "isced-97 code 4" into 2`
* `recode 5 "isced-97 code 5" into 3`
* `recode 6 "isced-97 code 6" into 3`
* `recode 95 "still in school" into 996`
* `recode 97 "other" into 999`
* `recode -2 "refusal" into 997`
* `recode -1 "don't know" into 998`
* `missing values into previous waves education if possible, otherwise, if mn101_==1, into 996`

**R script:**

```{r harmo4,   echo=TRUE}


SHARE_w4_educ <- left_join(SHARE_w4_educ,SHARE_ds_w1, by="id")
SHARE_w4_educ <- left_join(SHARE_w4_educ,SHARE_ds_w2, by="id", suffix = c("_w1","_w2"))
SHARE_ds_w4 <- tibble(id=SHARE_w4_educ$id)
SHARE_ds_w4$education <- car::recode(SHARE_w4_educ$isced_r, "3=2; 4=2; 5=3; 6=3; 95=996; 97=999;-2=997; -1=998;  NA=999")
SHARE_ds_w4$education[which(SHARE_w4_educ$mn101_==1)] <- 996
SHARE_ds_w4$education[which(is.na(SHARE_w4_educ$isced_r))] <- car::recode(SHARE_w4_educ$education_w2[which(is.na(SHARE_w4_educ$isced_r))],"NA=999")
SHARE_ds_w4$education[which(is.na(SHARE_w4_educ$isced_r) & SHARE_ds_w4$education==999)] <- car::recode(SHARE_w4_educ$education_w1[which(is.na(SHARE_w4_educ$isced_r) & SHARE_ds_w4$education==999)],"NA=999")




```

#### Statistical description of the new harmonised variable
```{r descript4,  echo=FALSE}
vari <- SHARE_ds_w4$education
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w4, aes(x=factor(education))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Level of education") + ylab("Frequency")
```

#### Validation
```{r crosstabulation4,  echo=FALSE}

```

 
 

 
```{r importOPAL4,  echo=FALSE}
label(SHARE_ds_w4$education) <- ds_label
education <- SHARE_ds_w4
save(education, file = "../RData/w4/education.RData")
#opal.assign.data(o, 'SHARE_ds_w4', SHARE_ds_w4)
#opal.symbol_import( o,'SHARE_ds_w4', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w4 <- SHARE_ds_w4
SHARE_m_ds_w4$education <- car::recode(SHARE_m_ds_w4$education, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w4', SHARE_m_ds_w4)
#opal.symbol_import( o,'SHARE_m_ds_w4', project='_Athlos_Harm_Dataset')
```







### Wave 5 

#### Study-specific variable description


| **Name** | `isced2011_r` |`mn101_`|
|----------|-------|-----|
| **Label** | `ISCED-11 coding of education` |`Interview version`|
| **Table name** | `sharew5_rel1-0-0_gv_isced` |`sharew5_rel1-0-0_dn`|
| **Categories** | `0 = none`<br/>`1 = isced-11 code 1`<br/>`2 = isced-11 code 2`<br/>`3 = isced-11 code 3`<br/>`4 = isced-11 code 4`<br/>`5 = isced-11 code 5`<br/>`6 = isced-11 code 6`<br/>`7 = ISCED-11 code 7`<br/>`8 = ISCED-11 code 8`<br/>`95 = still in school`<br/>`97 = other` |`0 = Baseline questionnaire`<br/>`1 = Longitudinal questionnaire`|
| **Missings** | `-2 = refusal`<br/>`-1 = don't know` |`NA`|
| **Description** |  |  |



```{r assign5,  echo=FALSE}

opal.assign.table.tibble(o, 'SHARE_w5_educ','SHARE.sharew5_rel1-0-0_gv_isced',variables=list('isced2011_r'), missings = TRUE)
SHARE_w5_educ <- opal.execute(o,'SHARE_w5_educ')
opal.assign.table.tibble(o, 'SHARE_w5_id','SHARE.sharew5_rel1-0-0_cv_r_without_noint', variables=list('hhid5'), missings = TRUE)
SHARE_w5_id <- opal.execute(o,'SHARE_w5_id')
opal.assign.table.tibble(o, 'SHARE_w5','SHARE.sharew5_rel1-0-0_dn', variables=list('mn101_'), missings = TRUE)
SHARE_w5 <- opal.execute(o,'SHARE_w5')

SHARE_w5_educ <- left_join(SHARE_w5_educ,SHARE_w5, by = "id")
SHARE_w5_educ <- left_join(SHARE_w5_id,SHARE_w5_educ,by="id")

```

```{r local5,  echo=FALSE}


vari <- SHARE_w5_educ$isced2011_r
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "isced2011_r") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-2,-1))[2], caption = "Type of missing")
ggplot(SHARE_w5_educ, aes(x=factor(isced2011_r))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Level of education") + ylab("Frequency")

vari <- SHARE_w5_educ$mn101_
kable(Categorical_summary(vari, missing_values = c(-2,-1))[3], caption = "mn101_") 
kable(Categorical_summary(vari, missing_values = c(-2,-1))[1], caption = "Category")
ggplot(SHARE_w5_educ, aes(x=factor(mn101_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vari)) + ylab("Frequency")

kable(table(SHARE_w5_educ$isced2011_r,SHARE_w5_educ$mn101_,useNA = "ifany"),caption = 'Variable mn101_ produces a skip')

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep "None" into 0`
* `keep "ISCED-11 code 1" into 1`
* `keep "ISCED-11 code 2" into 2`
* `recode 3 "ISCED-11 code 3" into 2`
* `recode 4 "ISCED-11 code 4" into 2`
* `recode 5 "ISCED-11 code 5" into 3`
* `recode 6 "ISCED-11 code 6" into 3`
* `recode 7 "ISCED-11 code 7" into 3`
* `recode 8 "ISCED-11 code 8" into 3`
* `recode 95 "still in school" into 996`
* `recode 97 "other" into 999`
* `recode -2 "refusal" into 997`
* `recode -1 "don't know" into 998`
* `missing values into previous waves education if possible, otherwise, if mn101_==1, into 996`

**R script:**

```{r harmo5,   echo=TRUE}



SHARE_w5_educ <- left_join(SHARE_w5_educ,SHARE_ds_w1, by="id")
SHARE_w5_educ <- left_join(SHARE_w5_educ,SHARE_ds_w2, by="id", suffix = c("_w1","_w2"))
SHARE_w5_educ <- left_join(SHARE_w5_educ,SHARE_ds_w4, by="id")
SHARE_ds_w5 <- tibble(id=SHARE_w5_educ$id)
SHARE_ds_w5$education <- car::recode(SHARE_w5_educ$isced2011_r, "3=2; 4=2; 5=3; 6=3;7=3; 8=3; 95=996; 97=999;-2=997; -1=998;  NA=999")
SHARE_ds_w5$education[which(SHARE_w5_educ$mn101_==1)] <- 996
SHARE_ds_w5$education[which(is.na(SHARE_w5_educ$isced2011_r))] <- car::recode(SHARE_w5_educ$education[which(is.na(SHARE_w5_educ$isced2011_r))],"NA=999")
SHARE_ds_w5$education[which(is.na(SHARE_w5_educ$isced2011_r) & SHARE_ds_w5$education==999)] <- car::recode(SHARE_w5_educ$education_w2[which(is.na(SHARE_w5_educ$isced2011_r) & SHARE_ds_w5$education==999)],"NA=999")
SHARE_ds_w5$education[which(is.na(SHARE_w5_educ$isced2011_r) & SHARE_ds_w5$education==999)] <- car::recode(SHARE_w5_educ$education_w1[which(is.na(SHARE_w5_educ$isced2011_r) & SHARE_ds_w5$education==999)],"NA=999")

```

#### Statistical description of the new harmonised variable
```{r descript5,  echo=FALSE}
vari <- SHARE_ds_w5$education
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w5, aes(x=factor(education))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Level of education") + ylab("Frequency")
```

#### Validation
```{r crosstabulation5,  echo=FALSE}

```

 
 

 
```{r importOPAL5,  echo=FALSE}
label(SHARE_ds_w5$education) <- ds_label
education <- SHARE_ds_w5
save(education, file = "../RData/w5/education.RData")
#opal.assign.data(o, 'SHARE_ds_w5', SHARE_ds_w5)
#opal.symbol_import( o,'SHARE_ds_w5', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w5 <- SHARE_ds_w5
SHARE_m_ds_w5$education <- car::recode(SHARE_m_ds_w5$education, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w5', SHARE_m_ds_w5)
#opal.symbol_import( o,'SHARE_m_ds_w5', project='_Athlos_Harm_Dataset')
```






## Summary of descriptive statistics of the harmonised variable accross populations and waves

Two tables are generated: 

1. Percentages of categories in each harmonised variable.

2. Frequencies and percentages of individuals with different trajectories.


```{r helpfunctions, echo=F}


labelling <- function(l.hds,m.hds){
  
  # Labelling of the tibbles with categorical data and creating new tibbles with all missings recodified as NA

  for(name in names(l.hds)) {
    # In the aux_object we copy the old tibble to recodify all missing values as NA.
    aux_object <- l.hds[[name]]
    # Labelling of variables
    label(l.hds[[name]][[2]]) <- label(aux_object[[2]]) <- ds_label
    # Labelling of categories (for continues variables, only missing values)
    l.hds[[name]][[2]] <- labelled(l.hds[[name]][[2]], labels = cat_label)
    aux_object[[2]] <- car::recode(aux_object[[2]], "miss_values_vector=NA")
    # Labelling of categories (for categorical variables, only non-missing values)
    aux_object[[2]] <- labelled(aux_object[[2]], labels = cat_label[1:4])
  # Saving the recodified tibble in list m.hds
    m.hds[[name]] <- aux_object
    rm(aux_object)
  }
  return(list(l.hds,m.hds))

}

# Creation of summary tables for categorical data

summaries <- function(l.hds,m.hds,lnames){

  # Creation of columns with categories and labels
  t.hds <- frq(l.hds[[1]][2])[[1]][,c(1,2)] 
  # For each wave/population in l.hds, add the correponding values
  for (i in seq_along(l.hds)){
    t.hds[2+i] <- frq(l.hds[[i]][2])[[1]][,4] 
  }
  # Add sample size for each wave/population
  t.hds[15,] <- c("n", "sample size", sapply(l.hds,function(wave) length(wave[[1]]))
  )
  # Add wave/population names
  names(t.hds)<-c("val", "label",lnames)
  return(t.hds)
  
}

# Creation of trajectories table for each population

trajectories <- function(m.hds,vbl_name){
  
  # First wave data
  dbb <- m.hds[[1]][,c("id",vbl_name)]
  # Merge with next waves data
  for(ind in 2:length(m.hds)){
    dbb <- merge(dbb, m.hds[[ind]][,c("id",vbl_name)], by = "id", suffixes = c("", paste0(".",names(m.hds)[ind])), all = T)
  }
  names(dbb) <- c("id", names(m.hds))
  
  # Glue individual data through all waves into trajectories
  v.dbb <- dbb[,2]
  for(ind in 2:length(m.hds)){
    v.dbb <- paste(v.dbb,dbb[,ind+1],sep="")
  }
  # Trajectories and frequencies
  f.dbb <- frq(v.dbb)[[1]][,c(1,2,4)]
  return(f.dbb)
  
}

# Save data tables

savingRD <- function(l.hds,vbl_name){
  
  for(index in seq_along(l.hds)){
    assign(vbl_name,l.hds[[index]])
    save(vbl_name,list = vbl_name, file = paste0(datafolder,names(l.hds)[index],"/",vbl_name,".RData"))
  }

}

```




```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names



share.cw <- list(w1 = c("w1","W1"), w2 = c("w2","W2"), w3 = c("w3","W3"), w4 = c("w4","W4"), w5 = c("w5","W5"))

# Consider only harmonised waves
l.hds <- lapply(share.cw, function(wname) if(exists(paste0("SHARE_ds_",wname[1]))){wname = list(get(paste0("SHARE_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing table of categories
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Printing table of trajectories
  f.dbb <- trajectories(lmlist[[2]],vbl_name=hd_vbl)
  print(knitr::kable(f.dbb))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```



# Quality estimation

The SHARE study used the ISCED to create the education variable. They used the ISCED-1997 in w1, w2, w4 and ISCED-2011 in w5. We transformed that into a 4-label categorical variable taking into account the ISCED directions. 

From SHARE Release Guide 6.1.1, section 14.4 gv_isced:

> Education is one of the most diverse international variables. SHARE is using the International Standard Classification of Education (ISCED) which allows for the standardised reporting of education statistics according to an internationally agreed set of definitions and concepts (for further information see http://uis.unesco.org/en/topic/international-standard-classification-education-isced). The gv_isced module contains the 1997 International Standard Classification of Education (ISCED-97). It is not only provided for respondents´ educational level but also for respondents´ children and former spouses´ as well as interviewers (the latter only in wave 1). In wave 1 and 2 the education of only up to four selected children was asked. In 2011, a revision to ISCED was adopted by the UNESCO Member States. From wave 5 onwards both ISCED versions (97 and 2011) are provided in gv_isced. Furthermore, waves 5 and 6 also contain the educational level of the respondents´ parents. Please note that we collect information on education only in the baseline interview of a respondent indicated by the variable mn101_ in the technical variables module. For the respondents´ children information on educational degrees is updated up to a certain age.

> The main difference between release 6.1.0 and 6.1.1 is related to the gv_isced module. In Switzerland, open answers of the education related questions in the DN module of wave 1 were used to supplement the variables isced1997_r and isced1997_sp. Further changes and corrections have been necessary in other countries and waves, too. Please note that due to the transfer of information from the baseline interview this also affects the gv_isced values of subsequent waves.

```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
