==> SHARE/1_Sociodemo_econom/6_living_alone.Rmd <==
---
title: "Living alone"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---
  
<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```
<!-- ########################################################## --> 

# Description of DataSchema variable

The description of harmonised variable is the following:


* Short name: `living_alone`
* Variable label: `Participant is currently living alone`
* Variable description: `Participant is currently living alone`
* Domain: `Socio-demographic and economic characteristics`
* Value type: `categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
0        | No
1        | Yes

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing



```{r lab, echo=FALSE}
ds_label <- "Living alone"
ds_label_all <- "Participant is currently living alone"
```

# Data process 

## SHARE

### Wave 1 

#### Study-specific variable description



| **Name** | `hhsize`|
|-|-|
| **Label** | `household size`|
| **Table name** | `sharew1_rel2-6-0_cv_r_whitout_nointw1` |
| **Description** |  |


```{r assign1, echo=F}
opal.assign.table.tibble(o, 'SHARE_w1_la','SHARE.sharew1_rel2-6-0_cv_r_whitout_nointw1',
                         variables=list('hhsize'), missings = TRUE)
SHARE_w1_la <- opal.execute(o,'SHARE_w1_la')
```

```{r local1, echo=F}
vari <- SHARE_w1_la$hhsize
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_w1_la, aes(x=factor(hhsize))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Living Alone") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep 1 into 1`
* `>1 into 0`
* `NA into 999`

**R script:**

```{r harmo1}
SHARE_ds_w1 <- tibble(id=SHARE_w1_la$id)
SHARE_ds_w1$living_alone <- car::recode(SHARE_w1_la$hhsize, "2:hi=0 ;NA=999")
SHARE_ds_w1$living_alone <- labelled(SHARE_ds_w1$living_alone, labels = c("No"=0, "Yes"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1, echo=F}
vari <- SHARE_ds_w1$living_alone
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w1, aes(x=factor(living_alone))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Living Alone") + ylab("Frequency")
```

#### Validation
```{r crosstabulation1, echo=F}
BeforeH <- table(SHARE_w1_la$hhsize)
BeforeH <- c(sum(BeforeH[2:length(BeforeH)]), BeforeH[1])
AfterH <- table(SHARE_ds_w1$living_alone)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c(">=2->0","1")
kable(C)
```





```{r importOPAL1, echo=F}
label(SHARE_ds_w1$living_alone) <- ds_label_all
living_alone <- SHARE_ds_w1
save(living_alone, file = "../RData/w1/living_alone.RData")
#opal.assign.data(o, 'SHARE_ds_w1', SHARE_ds_w1)
#opal.symbol_import(o,'SHARE_ds_w1', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w1 <- SHARE_ds_w1
SHARE_m_ds_w1$living_alone <- car::recode(SHARE_m_ds_w1$living_alone, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w1', SHARE_m_ds_w1)
#opal.symbol_import(o,'SHARE_m_ds_w1', project='_Athlos_Harm_Dataset')
```





### Wave 2 

#### Study-specific variable description



| **Name** | `hhsize`|
|-|-|
| **Label** | `household size`|
| **Table name** | `sharew2_rel2-6-0_cv_r_without_noint` |
| **Description** |  |


```{r assign2, echo=F}
opal.assign.table.tibble(o, 'SHARE_w2_la','SHARE.sharew2_rel2-6-0_cv_r_without_noint',
                         variables=list('hhsize'), missings = TRUE)
SHARE_w2_la <- opal.execute(o,'SHARE_w2_la')
```

```{r local2, echo=F}
vari <- SHARE_w2_la$hhsize
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_w2_la, aes(x=factor(hhsize))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Living Alone") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `keep 1 into 1`
* `>1 into 0`
* `NA into 999`

**R script:**

```{r harmo2}
SHARE_ds_w2 <- tibble(id=SHARE_w2_la$id)
SHARE_ds_w2$living_alone <- car::recode(SHARE_w2_la$hhsize, "2:hi=0 ;NA=999")
SHARE_ds_w2$living_alone <- labelled(SHARE_ds_w2$living_alone, labels = c("No"=0, "Yes"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2, echo=F}
vari <- SHARE_ds_w2$living_alone
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w2, aes(x=factor(living_alone))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Living Alone") + ylab("Frequency")
```

#### Validation
```{r crosstabulation2, echo=F}
BeforeH <- table(SHARE_w2_la$hhsize)
BeforeH <- c(sum(BeforeH[2:length(BeforeH)]), BeforeH[1])
AfterH <- table(SHARE_ds_w2$living_alone)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c(">=2->0","1")
kable(C)
```





```{r importOPAL2, echo=F}
label(SHARE_ds_w2$living_alone) <- ds_label_all
living_alone <- SHARE_ds_w2
save(living_alone, file = "../RData/w2/living_alone.RData")
#opal.assign.data(o, 'SHARE_ds_w2', SHARE_ds_w2)
#opal.symbol_import(o,'SHARE_ds_w2', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w2 <- SHARE_ds_w2
SHARE_m_ds_w2$living_alone <- car::recode(SHARE_m_ds_w2$living_alone, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w2', SHARE_m_ds_w2)
#opal.symbol_import(o,'SHARE_m_ds_w2', project='_Athlos_Harm_Dataset')
```




### Wave 3 

#### Study-specific variable description



| **Name** | `hhsize`|
|-|-|
| **Label** | `household size`|
| **Table name** | `sharew3_rel1_cv_r_without_noint` |
| **Description** |  |

```{r assign3, echo=F}
opal.assign.table.tibble(o, 'SHARE_w3_la','SHARE.sharew3_rel1_cv_r_without_noint',
                         variables=list('hhsize'), missings = TRUE)
SHARE_w3_la <- opal.execute(o,'SHARE_w3_la')
```

```{r local3, echo=F}
vari <- SHARE_w3_la$hhsize
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_w3_la, aes(x=factor(hhsize))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Living Alone") + ylab("Frequency")
```

There are values 0, they are to be considered as missing since at least 1  person (the respondent) should be counted here. Therefore they are to be recoded as 999.

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `recode 0 into 999` 
* `keep 1 into 1`
* `>1 into 0`
* `NA into 999`

**R script:**

```{r harmo3}
SHARE_ds_w3 <- tibble(id=SHARE_w3_la$id)
SHARE_ds_w3$living_alone <- car::recode(SHARE_w3_la$hhsize, "0=999; 2:hi=0;  NA=999")
SHARE_ds_w3$living_alone <- labelled(SHARE_ds_w3$living_alone, labels = c("No"=0, "Yes"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript3, echo=F}
vari <- SHARE_ds_w3$living_alone
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w3, aes(x=factor(living_alone))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Living Alone") + ylab("Frequency")
```

#### Validation
```{r crosstabulation3, echo=F}
BeforeH <- table(SHARE_w3_la$hhsize)
BeforeH <- c(sum(BeforeH[3:length(BeforeH)]), BeforeH[2], BeforeH[1])
AfterH <- table(SHARE_ds_w3$living_alone)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c(">=2->0","1", "(NA,0)->999")
kable(C)
```





```{r importOPAL3, echo=F}
label(SHARE_ds_w3$living_alone) <- ds_label_all
living_alone <- SHARE_ds_w3
save(living_alone, file = "../RData/w3/living_alone.RData")
#opal.assign.data(o, 'SHARE_ds_w3', SHARE_ds_w3)
#opal.symbol_import(o,'SHARE_ds_w3', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w3 <- SHARE_ds_w3
SHARE_m_ds_w3$living_alone <- car::recode(SHARE_m_ds_w3$living_alone, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w3', SHARE_m_ds_w3)
#opal.symbol_import(o,'SHARE_m_ds_w3', project='_Athlos_Harm_Dataset')
```






### Wave 4 

#### Study-specific variable description



| **Name** | `hhsize`|
|-|-|
| **Label** | `household size`|
| **Table name** | `sharew4_rel1-1-1_cv_r_without_noint` |
| **Description** |  |

```{r assign4, echo=F}
opal.assign.table.tibble(o, 'SHARE_w4_la','SHARE.sharew4_rel1-1-1_cv_r_without_noint',
                         variables=list('hhsize'), missings = TRUE)
SHARE_w4_la <- opal.execute(o,'SHARE_w4_la')
```

```{r local4, echo=F}
vari <- SHARE_w4_la$hhsize
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_w4_la, aes(x=factor(hhsize))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Living Alone") + ylab("Frequency")
```

There are values 0, they are to be considered as missing since at least 1  person (the respondent) should be counted here. Therefore they are to be recoded as 999.

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `recode 0 into 999` 
* `keep 1 into 1`
* `>1 into 0`
* `NA into 999`

**R script:**

```{r harmo4}
SHARE_ds_w4 <- tibble(id=SHARE_w4_la$id)
SHARE_ds_w4$living_alone <- car::recode(SHARE_w4_la$hhsize, "2:hi=0 ;c(0,NA)=999")
SHARE_ds_w4$living_alone <- labelled(SHARE_ds_w4$living_alone, labels = c("No"=0, "Yes"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript4, echo=F}
vari <- SHARE_ds_w4$living_alone
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w4, aes(x=factor(living_alone))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Living Alone") + ylab("Frequency")
```

#### Validation
```{r crosstabulation4, echo=F}
BeforeH <- table(SHARE_w4_la$hhsize)
BeforeH <- c(sum(BeforeH[3:length(BeforeH)]), BeforeH[2], BeforeH[1])
AfterH <- table(SHARE_ds_w4$living_alone)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c(">=2->0","1", "(NA,0)->999")
kable(C)
```





```{r importOPAL4, echo=F}
label(SHARE_ds_w4$living_alone) <- ds_label_all
living_alone <- SHARE_ds_w4
save(living_alone, file = "../RData/w4/living_alone.RData")
#opal.assign.data(o, 'SHARE_ds_w4', SHARE_ds_w4)
#opal.symbol_import(o,'SHARE_ds_w4', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w4 <- SHARE_ds_w4
SHARE_m_ds_w4$living_alone <- car::recode(SHARE_m_ds_w4$living_alone, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w4', SHARE_m_ds_w4)
#opal.symbol_import(o,'SHARE_m_ds_w4', project='_Athlos_Harm_Dataset')
```





### Wave 5 

#### Study-specific variable description



| **Name** | `hhsize`|
|-|-|
| **Label** | `household size`|
| **Table name** | `sharew5_rel1-0-0_cv_r_without_noint` |
| **Description** |  |

```{r assign5, echo=F}
opal.assign.table.tibble(o, 'SHARE_w5_la','SHARE.sharew5_rel1-0-0_cv_r_without_noint',
                         variables=list('hhsize'), missings = TRUE)
SHARE_w5_la <- opal.execute(o,'SHARE_w5_la')
```

```{r local5, echo=F}
vari <- SHARE_w5_la$hhsize
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_w5_la, aes(x=factor(hhsize))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Living Alone") + ylab("Frequency")
```

There are values 0, they are to be considered as missing since at least 1  person (the respondent) should be counted here. Therefore they are to be recoded as 999.

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `recode 0 into 999` 
* `keep 1 into 1`
* `>1 into 0`
* `NA into 999`

**R script:**

```{r harmo5}
SHARE_ds_w5 <- tibble(id=SHARE_w5_la$id)
SHARE_ds_w5$living_alone <- car::recode(SHARE_w5_la$hhsize, "2:hi=0; c(0,NA)=999")
SHARE_ds_w5$living_alone <- labelled(SHARE_ds_w5$living_alone, labels = c("No"=0, "Yes"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript5, echo=F}
vari <- SHARE_ds_w5$living_alone
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w5, aes(x=factor(living_alone))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Living Alone") + ylab("Frequency")
```

#### Validation
```{r crosstabulation5, echo=F}
BeforeH <- table(SHARE_w5_la$hhsize)
BeforeH <- c(sum(BeforeH[3:length(BeforeH)]), BeforeH[2], BeforeH[1])
AfterH <- table(SHARE_ds_w5$living_alone)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c(">=2->0","1", "(NA,0)->999")
kable(C)
```





```{r importOPAL5, echo=F}
label(SHARE_ds_w5$living_alone) <- ds_label_all
living_alone <- SHARE_ds_w5
save(living_alone, file = "../RData/w5/living_alone.RData")
#opal.assign.data(o, 'SHARE_ds_w5', SHARE_ds_w5)
#opal.symbol_import(o,'SHARE_ds_w5', project='_Athlos_Harm_Dataset')

SHARE_m_ds_w5 <- SHARE_ds_w5
SHARE_m_ds_w5$living_alone <- car::recode(SHARE_m_ds_w5$living_alone, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'SHARE_m_ds_w5', SHARE_m_ds_w5)
#opal.symbol_import(o,'SHARE_m_ds_w5', project='_Athlos_Harm_Dataset')
```




# Quality estimation
No comments.


```{r closeRsession, echo=FALSE}
opal.logout(o)
```
