==> SHARE/1_Sociodemo_econom/13_Wealth.Rmd <==
---
title: "Wealth"
output: 
  html_document:
    theme: united
    highlight: tango
    toc: true
    toc_depth: 6
    toc_float:
      collapsed: true
      smooth_scroll: true
    number_sections: true
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->



```{r setup, include=FALSE,  echo=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```

<!-- ########################################################## --> 

# Description of DataSchema variable 

* Short name: `wealth`
* Variable label: `Total score based in all sources of income transformed into a comparable and normalized categoric variable.`
* Domain: `Socio-demographic and economic characteristics`
* Value type: `categorical ordinal`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
1        | 1st quintile
2        | 2nd quintile
3        | 3rd quintile
4        | 4th quintile
5        | 5th quintile

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

* Comments: `Need to pay attention to specificities like: before/after taxes or contributions. People: HH income is prioritized. When not available individual/caregiver/informant/family income is caolected. Source: Pension/family/work/government/disablity/total. all sources are taken into account. For some studies, the maximum value can be compared (and crossed by sex and age to control biases) . Time-lapse: weekly, monthly, in the last 12 months. 'Many values are missing since not all respondents have all sources of income or, within the household there is not always more members, in these cases do not confuse missing data with missing values and mean a 0 should be added to the total sum in each case.`

```{r lab, echo=FALSE}
ds_label <- "Wealth"
ds_label_all <- "Total score based in all sources of income transformed into a comparable and normalized categoric variable"
```


# Data process


```{r, dataDirectory, echo=FALSE}
datafolder <- "../RData"
```

## SHARE

### Wave 1 

#### Study-specific variable description


| **Name** | `income_pct_w1`| 
|-|-|
| **Label** | `Hhd income percentiles wave 1` |
| **Table name** | `easySHARE_w1` |
| **Categories** | `continuous` |
| **Missing values** | `-13 = not asked in this wave` | 
| **Description** |  | 

```{r assign_aksst1, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SHARE/RData/w1/living_status.RData")

opal.assign.table.tibble(o, 'SHARE_wealth2','SHARE.easySHARE_w1',
                         variables=list('income_pct_w1'), missings = TRUE)
SHARE_wealth2 <- opal.execute(o,'SHARE_wealth2')

SHARE_wealth <- left_join(living_status, SHARE_wealth2, by='id')
```

```{r local_aksst1,echo=F}
vari <- SHARE_wealth$income_pct_w1
kable(Categorical_summary(vari, missing_values = c(-13))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-13))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-13))[2], caption = "Type of missing")
ggplot(SHARE_wealth, aes(x=factor(income_pct_w1))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonized variable from the study-scpecific variable it have to be recoded as follows: 

* `1 and 2 into 1`
* `3 and 4 into 2`
* `5 and 6 into 3`
* `7 and 8 into 4`
* `9 and 10 into 5`
* `-13 into 994`
* `NA as 999`

**R script:**

```{r harmo1}
SHARE_ds_w1 <- tibble(id=SHARE_wealth$id)
SHARE_ds_w1$wealth <- car::recode(SHARE_wealth$income_pct_w1, "2=1; c(3,4)=2; c(5,6)=3; c(7,8)=4; c(9,10)=5;  NA=999")
SHARE_ds_w1$wealth <- as.integer(SHARE_ds_w1$wealth)
SHARE_ds_w1$wealth <- labelled(SHARE_ds_w1$wealth, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript_aksst1,echo=F}
vari <- SHARE_ds_w1$wealth
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w1, aes(x=factor(wealth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Validation
```{r crosstabulation1, echo=FALSE}
```

```{r importOPAL1, echo=FALSE}
label(SHARE_ds_w1$wealth) <- ds_label_all
wealth <- SHARE_ds_w1
save(wealth, file = "../RData/w1/wealth.RData")
rm(wealth)
```


```{r read1, include=FALSE}
rm(SHARE_gs)
```


### Wave 2

#### Study-specific variable description


| **Name** | `income_pct_w2`| 
|-|-|
| **Label** | `Hhd income percentiles wave 2` |
| **Table name** | `easySHARE_income_w2` |
| **Categories** | `continuous` |
| **Missing values** | `-15 = no information` | 
| | `-14 = not asked in this country` |
| | `-13 = not asked in this wave` |
| | `-12 = don't know / refusal` |
| | `-7 = not yet coded` |
| | `-3 = implausible value/suspected wrong` |
| **Description** |  | 

    
```{r assign2, include=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SHARE/RData/w2/living_status.RData")

opal.assign.table.tibble(o, 'SHARE_wealth1','SHARE.easySHARE_income_w2',
                         variables=list('income_pct_w2'), missings = TRUE)
SHARE_wealth1 <- opal.execute(o,'SHARE_wealth1')

SHARE_wealth <- left_join(living_status, SHARE_wealth1, by='id')
```

```{r local2, echo=F}
vari <- SHARE_wealth$income_pct_w2
kable(Categorical_summary(vari, missing_values = c(-15, -14, -13, -12, -7, -3))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-15, -14, -13, -12, -7, -3))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-15, -14, -13, -12, -7, -3))[2], caption = "Type of missing")
ggplot(SHARE_wealth, aes(x=factor(income_pct_w2))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonized variable from the study-scpecific variable it have to be recoded as follows: 

* `1 and 2 into 1`
* `3 and 4 into 2`
* `5 and 6 into 3`
* `7 and 8 into 4`
* `9 and 10 into 5`
* `-15 into 999`
* `-14 into 994`
* `-13 into 994`
* `-12 into 999`
* `-7 into 994`
* `-3 into 991`
* `NA as 999`

**R script:**

```{r harmo2}
SHARE_ds_w2 <- tibble(id=SHARE_wealth$id)
SHARE_ds_w2$wealth <- car::recode(SHARE_wealth$income_pct_w2, "2=1; c(3,4)=2; c(5,6)=3; c(7,8)=4; c(9,10)=5; -15=999; -14=994; -13=994; -12=999; -7=994; -3=991;  NA=999")
SHARE_ds_w2$wealth <- as.integer(SHARE_ds_w2$wealth)
SHARE_ds_w2$wealth <- labelled(SHARE_ds_w2$wealth, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript2, echo=F}
vari <- SHARE_ds_w2$wealth
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w2, aes(x=factor(wealth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Validation
```{r crosstab2, echo=F}
```

```{r read2, include=FALSE}
label(SHARE_ds_w2$wealth) <- ds_label_all
wealth <- SHARE_ds_w2
save(wealth, file = "../RData/w2/wealth.RData")
rm(wealth)
```


### Wave 3


There is no variable. 


### Wave 4

#### Study-specific variable description


| **Name** | `income_pct_w4`| 
|-|-|
| **Label** | `Hhd income percentiles wave 4` |
| **Table name** | `easySHARE_income_w4` |
| **Categories** | `continuous` |
| **Missing values** | `-15 = no information` | 
| | `-14 = not asked in this country` |
| | `-13 = not asked in this wave` |
| | `-12 = don't know / refusal` |
| | `-7 = not yet coded` |
| | `-3 = implausible value/suspected wrong` |
| **Description** |  | 
    
```{r assign4, include=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SHARE/RData/w4/living_status.RData")
opal.assign.table.tibble(o, 'SHARE_wealth2','SHARE.easySHARE_income_w4',
                         variables=list('income_pct_w4'), missings = TRUE)
SHARE_wealth2 <- opal.execute(o,'SHARE_wealth2')
SHARE_wealth <- left_join(living_status, SHARE_wealth2, by='id')
```

```{r local4, echo=F}
vari <- SHARE_wealth$income_pct_w4
kable(Categorical_summary(vari, missing_values = c(-15, -14, -13, -12, -7, -3))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-15, -14, -13, -12, -7, -3))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-15, -14, -13, -12, -7, -3))[2], caption = "Type of missing")
ggplot(SHARE_wealth, aes(x=factor(income_pct_w4))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonized variable from the study-scpecific variable it have to be recoded as follows: 

* `1 and 2 into 1`
* `3 and 4 into 2`
* `5 and 6 into 3`
* `7 and 8 into 4`
* `9 and 10 into 5`
* `-15 into 999`
* `-14 into 994`
* `-13 into 994`
* `-12 into 999`
* `-7 into 994`
* `-3 into 991`
* `NA as 999`

**R script:**

```{r harmo4}
SHARE_ds_w4 <- tibble(id=SHARE_wealth$id)
SHARE_ds_w4$wealth <- car::recode(SHARE_wealth$income_pct_w4, "2=1; c(3,4)=2; c(5,6)=3; c(7,8)=4; c(9,10)=5; -15=999; -14=994; -13=994; -12=999; -7=994; -3=991;  NA=999")
SHARE_ds_w4$wealth <- as.integer(SHARE_ds_w4$wealth)
SHARE_ds_w4$wealth <- labelled(SHARE_ds_w4$wealth, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript4, echo=F}
vari <- SHARE_ds_w4$wealth
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w4, aes(x=factor(wealth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Validation
```{r crosstab4, echo=F}
```


```{r read4, include=FALSE}
label(SHARE_ds_w4$wealth) <- ds_label
wealth <- SHARE_ds_w4
save(wealth, file = "../RData/w4/wealth.RData")
rm(SHARE_gs)
```



### Wave 5

#### Study-specific variable description


| **Name** | `income_pct_w5`| 
|-|-|
| **Label** | `Hhd income percentiles wave 5` |
| **Table name** | `easySHARE_income_w5` |
| **Categories** | `continuous` |
| **Missing values** | `-15 = no information` | 
| | `-14 = not asked in this country` |
| | `-13 = not asked in this wave` |
| | `-12 = don't know / refusal` |
| | `-7 = not yet coded` |
| | `-3 = implausible value/suspected wrong` |
| **Description** |  | 

    
```{r assign5, include=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SHARE/RData/w5/living_status.RData")
opal.assign.table.tibble(o, 'SHARE_wealth2','SHARE.easySHARE_income_w5',
                         variables=list('income_pct_w5'), missings = TRUE)
SHARE_wealth2 <- opal.execute(o,'SHARE_wealth2')

SHARE_wealth <- left_join(living_status, SHARE_wealth2, by='id')
```

```{r local5, echo=F}
vari <- SHARE_wealth$income_pct_w5
kable(Categorical_summary(vari, missing_values = c(-15, -14, -13, -12, -7, -3))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(-15, -14, -13, -12, -7, -3))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(-15, -14, -13, -12, -7, -3))[2], caption = "Type of missing")
ggplot(SHARE_wealth, aes(x=factor(income_pct_w5))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonized variable from the study-scpecific variable it have to be recoded as follows: 

* `1 and 2 into 1`
* `3 and 4 into 2`
* `5 and 6 into 3`
* `7 and 8 into 4`
* `9 and 10 into 5`
* `-15 into 999`
* `-14 into 994`
* `-13 into 994`
* `-12 into 999`
* `-7 into 994`
* `-3 into 991`
* `NA as 999`

**R script:**

```{r harmo5}
SHARE_ds_w5 <- tibble(id=SHARE_wealth$id)
SHARE_ds_w5$wealth <- car::recode(SHARE_wealth$income_pct_w5, "2=1; c(3,4)=2; c(5,6)=3; c(7,8)=4; c(9,10)=5; -15=999; -14=994; -13=994; -12=999; -7=994; -3=991;  NA=999")
SHARE_ds_w5$wealth <- as.integer(SHARE_ds_w5$wealth)
SHARE_ds_w5$wealth <- labelled(SHARE_ds_w5$wealth, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript5, echo=F}
vari <- SHARE_ds_w5$wealth
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w5, aes(x=factor(wealth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Validation
```{r crosstab5, echo=F}
```

```{r read5, include=FALSE}
label(SHARE_ds_w5$wealth) <- ds_label
wealth <- SHARE_ds_w5
save(wealth, file = "../RData/w5/wealth.RData")
rm(SHARE_gs)
```



# Quality estimation
After taxes income from employment and pensions, household income, the housing, and assets form the whole household was included in this calculation.


<!--- #### Close OPAL R Session -->
```{r closeRsession, echo=FALSE}
opal.logout(o)
```
