==> SHARE/1_Sociodemo_econom/14_country.Rmd <==
---
title: "Country"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---
  
<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```
<!-- ########################################################## --> 


# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `country`
* Variable label: `country identifier`
* Variable description: `country identifier`
* Domain: `Socio-demographic and economic characteristics`
* Value type: `categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
51       | Austria
59       | Germany
71       | Sweden  
66       | Netherlands
70       | Spain
63       | Italy
58       | France
55       | Denmark
60       | Greece
72       | Switzerland
52       | Belgium
33       | Israel
54       | Czech Republic
65       | Luxembourg
67       | Poland
62       | Ireland
61       | Hungary
68       | Portugal
69       | Slovenia
56       | Estonia
53       | Croatia


* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Country"
ds_label_all <- "Country identifier"
```


# Data process

## SHARE

### Wave 1 

#### Study-specific variable description

| **Name** | `country` |
|-|-|
| **Label** | `country identifier` | 
| **Table name** | `sharew1_rel2-6-0_gv_weights` |
| **Categories** | `11 = Austria` | 
| | `12 = Germany` | 
| | `13 = Sweden` | 
| | `14 = Netherlands` | 
| | `15 = Spain` | 
| | `16 = Italy` | 
| | `17 = France` | 
| | `18 = Denmark` | 
| | `19 = Greece` | 
| | `20 = Switzerland` | 
| | `23 = Belgium` |
| | `25 = Israel` | 
| | `28 = Czechia` |
| | `29 = Poland` | 
| | `30 = Ireland` | 
| | `32 = Hungary` | 
| | `33 = Portugal` | 
| | `34 = Slovenia` | 
| | `35 = Estonia` | 
| **Description** |  | 

```{r assign1, echo=FALSE}
opal.assign.table.tibble(o, 'SHARE_country','SHARE.sharew1_rel2-6-0_gv_weights',
                         variables=list('country'), missings = TRUE)
SHARE_country <- opal.execute(o,'SHARE_country')
```

```{r local1, echo=FALSE}
vari <- SHARE_country$country
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_country, aes(x=factor(country))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Harmonisation algorithm

To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `recode 11 into 51`
* `recode 12 into 59`
* `recode 13 into 71`
* `recode 14 into 66`
* `recode 15 into 70`
* `recode 16 into 63`
* `recode 17 into 58`
* `recode 18 into 55`
* `recode 19 into 60`
* `recode 20 into 72`
* `recode 23 into 52`
* `recode 25 into 33`
* `recode 28 into 54`
* `recode 29 into 67`
* `recode 30 into 62`
* `recode 32 into 61`
* `recode 33 into 68`
* `recode 34 into 69`
* `recode 35 into 56`

**R script:**

```{r harmo1}
SHARE_ds_w1 <- tibble(id=SHARE_country$id)
SHARE_ds_w1$country <- car::recode(SHARE_country$country, "11=51; 12=59; 13=71; 14=66; 15=70; 16=63; 17=58; 18=55; 19=60; 20=72; 23=52; 25=33; 28=54; 29=67; 30=62; 32=61; 33=68; 34=69; 35=56")
SHARE_ds_w1$country <- labelled(SHARE_ds_w1$country, labels = c("Austria"=51, "Germany"=59, "Sweden"=71, "Netherlands"=66, "Spain"=70, "Italy"=63, "France"=58, "Denmark"=55, "Greece"=60, "Switzerland"=72, "Belgium"=52, "Israel"=33, "Czech Republic"=54, "Poland"=67, "Ireland"=62, "Hungary"=61, "Portugal"=68, "Slovenia"=69, "Estonia"=56,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```


#### Statistics of the new harmonised variable
```{r descript1, echo=F}
vari <- SHARE_ds_w1$country
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w1, aes(x=factor(country))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation1, echo=F,results='asis'}
```



```{r importOPAL1, echo=F}
label(SHARE_ds_w1$country) <- ds_label
country <- SHARE_ds_w1
save(country, file = "../RData/w1/country.RData")
#opal.assign.data(o, 'SHARE_ds_w1', SHARE_ds_w1)
#opal.symbol_import(o,'SHARE_ds_w1', project='_Athlos_Harm_Dataset')
```



### Wave 2 

#### Study-specific variable description

| **Name** | `country` |
|-|-|
| **Label** | `country identifier` | 
| **Table name** | `sharew2_rel2-6-0_cv_r_without_noint` |
| **Categories** | `11 = Austria` |
| | `12 = Germany` |
| | `13 = Sweden` |
| | `14 = Netherlands` |
| | `15 = Spain` |
| | `16 = Italy` |
| | `17 = France` |
| | `18 = Denmark` |
| | `19 = Greece` |
| | `20 = Switzerland` |
| | `23 = Belgium` |
| | `25 = Israel` |
| | `28 = Czechia` |
| | `29 = Poland` |
| | `30 = Ireland` |
| | `32 = Hungary` |
| | `33 = Portugal` |
| | `34 = Slovenia` |
| | `35 = Estonia` |
| **Description** |  | 


```{r assign2, echo=F}
opal.assign.table.tibble(o, 'SHARE_country1','SHARE.sharew2_rel2-6-0_cv_r_without_noint',
                         variables=list('country'), missings = TRUE)
SHARE_country1 <- opal.execute(o,'SHARE_country1')
opal.assign.table.tibble(o, 'SHARE_country2','SHARE.sharew2_rel2-6-0_xt',
                         variables=list('country'), missings = TRUE)
SHARE_country2 <- opal.execute(o,'SHARE_country2')

SHARE_country <- full_join(SHARE_country1, SHARE_country2, by='id')
SHARE_country$country <- SHARE_country$country.x
SHARE_country$country[is.na(SHARE_country$country)] <- SHARE_country$country.y[is.na(SHARE_country$country)]
```

```{r local2, echo=F}
vari <- SHARE_country$country
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_country, aes(x=factor(country))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Harmonisation algorithm

To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `recode 11 into 51`
* `recode 12 into 59`
* `recode 13 into 71`
* `recode 14 into 66`
* `recode 15 into 70`
* `recode 16 into 63`
* `recode 17 into 58`
* `recode 18 into 55`
* `recode 19 into 60`
* `recode 20 into 72`
* `recode 23 into 52`
* `recode 25 into 33`
* `recode 28 into 54`
* `recode 29 into 67`
* `recode 30 into 62`
* `recode 32 into 61`
* `recode 33 into 68`
* `recode 34 into 69`
* `recode 35 into 56`

**R script:**

```{r harmo2}
SHARE_ds_w2 <- tibble(id=SHARE_country$id)
SHARE_ds_w2$country <- car::recode(SHARE_country$country, "11=51; 12=59; 13=71; 14=66; 15=70; 16=63; 17=58; 18=55; 19=60; 20=72; 23=52; 25=33; 28=54; 29=67; 30=62; 32=61; 33=68; 34=69; 35=56")
SHARE_ds_w2$country <- labelled(SHARE_ds_w2$country, labels = c("Austria"=51, "Germany"=59, "Sweden"=71, "Netherlands"=66, "Spain"=70, "Italy"=63, "France"=58, "Denmark"=55, "Greece"=60, "Switzerland"=72, "Belgium"=52, "Israel"=33, "Czech Republic"=54, "Poland"=67, "Ireland"=62, "Hungary"=61, "Portugal"=68, "Slovenia"=69, "Estonia"=56,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2, echo=F}
vari <- SHARE_ds_w2$country
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w2, aes(x=factor(country))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation2, echo=F,results='asis'}
```


```{r importOPAL2, echo=F}
label(SHARE_ds_w2$country) <- ds_label
country <- SHARE_ds_w2
save(country, file = "../RData/w2/country.RData")
rm(country, SHARE_death2, SHARE_death2)
#opal.assign.data(o, 'SHARE_ds_w2', SHARE_ds_w2)
#opal.symbol_import(o,'SHARE_ds_w2', project='_Athlos_Harm_Dataset')
```




### Wave 3 

#### Study-specific variable description

| **Name** | `country` |
|-|-|
| **Label** | `country identifier` | 
| **Table name** | `sharew3_rel1_cv_r_without_noint` |
| **Categories** | `11 = Austria` |
| | `12 = Germany` |
| | `13 = Sweden` |
| | `14 = Netherlands` |
| | `15 = Spain` |
| | `16 = Italy` |
| | `17 = France` |
| | `18 = Denmark` |
| | `19 = Greece` |
| | `20 = Switzerland` |
| | `23 = Belgium` |
| | `25 = Israel` |
| | `28 = Czechia` |
| | `29 = Poland` |
| | `30 = Ireland` |
| | `32 = Hungary` |
| | `33 = Portugal` |
| | `34 = Slovenia` |
| | `35 = Estonia` |
| **Description** |  | 



```{r assign3, echo=F}
opal.assign.table.tibble(o, 'SHARE_country1','SHARE.sharew3_rel1_cv_r_without_noint',
                         variables=list('country'), missings = TRUE)
SHARE_country1 <- opal.execute(o,'SHARE_country1')
opal.assign.table.tibble(o, 'SHARE_country2','SHARE.sharew3_rel1_xt',
                         variables=list('country'), missings = TRUE)
SHARE_country2 <- opal.execute(o,'SHARE_country2')
opal.assign.table.tibble(o, 'SHARE_country3','SHARE.sharew3_rel1_gv_weights',
                         variables=list('country'), missings = TRUE)
SHARE_country3 <- opal.execute(o,'SHARE_country3')

names(SHARE_country3) <- c("id", "country.3")

SHARE_country <- full_join(SHARE_country1, SHARE_country2, by='id')
SHARE_country <- full_join(SHARE_country, SHARE_country3, by='id')

SHARE_country$country <- SHARE_country$country.x
SHARE_country$country[is.na(SHARE_country$country)] <- SHARE_country$country.y[is.na(SHARE_country$country)]
SHARE_country$country[is.na(SHARE_country$country)] <- SHARE_country$country.3[is.na(SHARE_country$country)]
```

```{r local3, echo=F}
vari <- SHARE_country$country
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_country, aes(x=factor(country))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `recode 11 into 51`
* `recode 12 into 59`
* `recode 13 into 71`
* `recode 14 into 66`
* `recode 15 into 70`
* `recode 16 into 63`
* `recode 17 into 58`
* `recode 18 into 55`
* `recode 19 into 60`
* `recode 20 into 72`
* `recode 23 into 52`
* `recode 25 into 33`
* `recode 28 into 54`
* `recode 29 into 67`
* `recode 30 into 62`
* `recode 32 into 61`
* `recode 33 into 68`
* `recode 34 into 69`
* `recode 35 into 56`

**R script:**

```{r harmo3}
SHARE_ds_w3 <- tibble(id=SHARE_country$id)
SHARE_ds_w3$country <- car::recode(SHARE_country$country, "11=51; 12=59; 13=71; 14=66; 15=70; 16=63; 17=58; 18=55; 19=60; 20=72; 23=52; 25=33; 28=54; 29=67; 30=62; 32=61; 33=68; 34=69; 35=56")
SHARE_ds_w3$country <- labelled(SHARE_ds_w3$country, labels = c("Austria"=51, "Germany"=59, "Sweden"=71, "Netherlands"=66, "Spain"=70, "Italy"=63, "France"=58, "Denmark"=55, "Greece"=60, "Switzerland"=72, "Belgium"=52, "Israel"=33, "Czech Republic"=54, "Poland"=67, "Ireland"=62, "Hungary"=61, "Portugal"=68, "Slovenia"=69, "Estonia"=56,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript3, echo=F}
vari <- SHARE_ds_w3$country
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w3, aes(x=factor(country))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation3, echo=F,results='asis'}
```



```{r importOPAL3, echo=F}
label(SHARE_ds_w3$country) <- ds_label
country <- SHARE_ds_w3
save(country, file = "../RData/w3/country.RData")
#opal.assign.data(o, 'SHARE_ds_w3', SHARE_ds_w3)
#opal.symbol_import(o,'SHARE_ds_w3', project='_Athlos_Harm_Dataset')
```




### Wave 4 

#### Study-specific variable description

| **Name** | `country` |
|-|-|
| **Label** | `country identifier` | 
| **Table name** | `sharew1_rel2-6-0_gv_weights` |
| **Categories** | `11 = Austria` |
| | `12 = Germany` |
| | `13 = Sweden` |
| | `14 = Netherlands` |
| | `15 = Spain` |
| | `16 = Italy` |
| | `17 = France` |
| | `18 = Denmark` |
| | `19 = Greece` |
| | `20 = Switzerland` |
| | `23 = Belgium` |
| | `25 = Israel` |
| | `28 = Czechia` |
| | `29 = Poland` |
| | `30 = Ireland` |
| | `32 = Hungary` |
| | `33 = Portugal` |
| | `34 = Slovenia` |
| | `35 = Estonia` |
| **Description** |  | 



```{r assign4, echo=F}
opal.assign.table.tibble(o, 'SHARE_country','SHARE.sharew4_rel1-1-1_cv_r_without_noint',
                         variables=list('country'), missings = TRUE)
SHARE_country <- opal.execute(o,'SHARE_country')
```

```{r local4, echo=F}
vari <- SHARE_country$country
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_country, aes(x=factor(country))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `recode 11 into 51`
* `recode 12 into 59`
* `recode 13 into 71`
* `recode 14 into 66`
* `recode 15 into 70`
* `recode 16 into 63`
* `recode 17 into 58`
* `recode 18 into 55`
* `recode 19 into 60`
* `recode 20 into 72`
* `recode 23 into 52`
* `recode 25 into 33`
* `recode 28 into 54`
* `recode 29 into 67`
* `recode 30 into 62`
* `recode 32 into 61`
* `recode 33 into 68`
* `recode 34 into 69`
* `recode 35 into 56`

**R script:**

```{r harmo4}
SHARE_ds_w4 <- tibble(id=SHARE_country$id)
SHARE_ds_w4$country <- car::recode(SHARE_country$country, "11=51; 12=59; 13=71; 14=66; 15=70; 16=63; 17=58; 18=55; 19=60; 20=72; 23=52; 25=33; 28=54; 29=67; 30=62; 32=61; 33=68; 34=69; 35=56")
SHARE_ds_w4$country <- labelled(SHARE_ds_w4$country, labels = c("Austria"=51, "Germany"=59, "Sweden"=71, "Netherlands"=66, "Spain"=70, "Italy"=63, "France"=58, "Denmark"=55, "Greece"=60, "Switzerland"=72, "Belgium"=52, "Israel"=33, "Czech Republic"=54, "Poland"=67, "Ireland"=62, "Hungary"=61, "Portugal"=68, "Slovenia"=69, "Estonia"=56,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript4, echo=F}
vari <- SHARE_ds_w4$country
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w4, aes(x=factor(country))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation4, echo=F,results='asis'}

```




```{r importOPAL4, echo=F}
label(SHARE_ds_w4$country) <- ds_label
country <- SHARE_ds_w4
save(country, file = "../RData/w4/country.RData")
#opal.assign.data(o, 'SHARE_ds_w4', SHARE_ds_w4)
#opal.symbol_import(o,'SHARE_ds_w4', project='_Athlos_Harm_Dataset')
```







### Wave 5 

#### Study-specific variable description

| **Name** | `country` |
|-|-|
| **Label** | `country identifier` | 
| **Table name** | `sharew1_rel2-6-0_gv_weights` |
| **Categories** | `11 = Austria` |
| | `12 = Germany` |
| | `13 = Sweden` |
| | `14 = Netherlands` |
| | `15 = Spain` |
| | `16 = Italy` |
| | `17 = France` |
| | `18 = Denmark` |
| | `19 = Greece` |
| | `20 = Switzerland` |
| | `23 = Belgium` |
| | `25 = Israel` |
| | `28 = Czechia` |
| | `29 = Poland` |
| | `30 = Ireland` |
| | `31 = Luxembourg` |
| | `32 = Hungary` |
| | `33 = Portugal` |
| | `34 = Slovenia` |
| | `35 = Estonia` |
| | `47 = Croatia` |
| **Description** |  | 



```{r assign5, echo=F}
opal.assign.table.tibble(o, 'SHARE_country','SHARE.sharew5_rel1-0-0_cv_r_without_noint',
                         variables=list('country'), missings = TRUE)
SHARE_country <- opal.execute(o,'SHARE_country')
```

```{r local5, echo=F}
vari <- SHARE_country$country
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(SHARE_country, aes(x=factor(country))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `recode 11 into 51`
* `recode 12 into 59`
* `recode 13 into 71`
* `recode 14 into 66`
* `recode 15 into 70`
* `recode 16 into 63`
* `recode 17 into 58`
* `recode 18 into 55`
* `recode 19 into 60`
* `recode 20 into 72`
* `recode 23 into 52`
* `recode 25 into 33`
* `recode 28 into 54`
* `recode 29 into 67`
* `recode 30 into 62`
* `recode 31 into 65`
* `recode 32 into 61`
* `recode 33 into 68`
* `recode 34 into 69`
* `recode 35 into 56`
* `recode 47 into 53`


**R script:**

```{r harmo5}
SHARE_ds_w5 <- tibble(id=SHARE_country$id)
SHARE_ds_w5$country <- car::recode(SHARE_country$country, "11=51; 12=59; 13=71; 14=66; 15=70; 16=63; 17=58; 18=55; 19=60; 20=72; 23=52; 25=33; 28=54; 29=67; 30=62; 32=61; 33=68; 34=69; 35=56; 31=65; 47=53")
SHARE_ds_w5$country <- labelled(SHARE_ds_w5$country, labels = c("Austria"=51, "Germany"=59, "Sweden"=71, "Netherlands"=66, "Spain"=70, "Italy"=63, "France"=58, "Denmark"=55, "Greece"=60, "Switzerland"=72, "Belgium"=52, "Israel"=33, "Czech Republic"=54, "Poland"=67, "Ireland"=62, "Hungary"=61, "Portugal"=68, "Slovenia"=69, "Estonia"=56,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript5, echo=F}
vari <- SHARE_ds_w5$country
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(SHARE_ds_w5, aes(x=factor(country))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation5, echo=F,results='asis'}
```





```{r importOPAL5, echo=F}
label(SHARE_ds_w5$country) <- ds_label
country <- SHARE_ds_w5
save(country, file = "../RData/w5/country.RData")
#opal.assign.data(o, 'SHARE_ds_w5', SHARE_ds_w5)
#opal.symbol_import(o,'SHARE_ds_w5', project='_Athlos_Harm_Dataset')
```



# Quality estimation


```{r closeRsession, echo=FALSE}
opal.logout(o)
```
