==> SHARE/1_Sociodemo_econom/15_migration.Rmd <==
---
title: "Migration"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/migrationDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/migrationDS.R')
```



# Data process

## SHARE

### Wave 1 

#### Study-specific variable description

| **Name** | `dn005c` | `dn004_` |
|-|-|-|
| **Label** | `other country of birth-country code` |`country of birth` |
| **Table name** | `sharew1_rel2-6-0_dn` |`sharew1_rel2-6-0_dn` |
| **Categories** | `2 = Africa` <br/>`54 = Afghanistan` <br/> `8 = Albania` <br/> `10 = Antarctica` <br/> `12 = Algeria` <br/> `16 = American Samoa` <br/> `20 = Andorra` <br/> `24 = Angola` <br/> `28 = Antigua and Barbuda` <br/> `31 = Azerbaijan` <br/> `32 = Argentina` <br/> `36 = Australia` <br/> `40 = Austria` <br/> `44 = Bahamas` <br/> `48 = Bahrain`  <br/> `50 = Bangladesh` <br/> `51 = Armenia` <br/> `52 = Barbados` <br/> `56 = Belgium` <br/> `60 = Bermuda` <br/> `64 = Bhutan` <br/> `68 = Bolivia` <br/> `70 = Bosnia and Herzegowina` <br/> `72 = Botswana` <br/> `74 = Bouvet Island` <br/> `76 = Brazil` <br/> `80 = British Antarctic Territory` <br/> `84 = Belize` <br/> `86 = British Indian Ocean Territory` <br/> `90 = Solomon Islands` <br/> `92 = Virgin Islands (british)` <br/> `96 = Brunei Darussalam` <br/> `100 = Bulgaria` <br/> `104 = Myanmar` <br/> `108 = Burundi` <br/> `112 = Belarus` <br/> `116 = Cambodia` <br/> `120 = Cameroon` <br/> `124 = Canada` <br/> `128 = Canton and Enderbury Islands` <br/> `132 = Cape Verde` <br/> `136 = Cayman Islands` <br/> `140 = Central African Republic` <br/> `144 = Sri Lanka` <br/> `148 = Chad` <br/> `152 = Chile ` <br/> `...` | `1 = yes` <br/> `5 = no` |
| **Missings** | `NA`<br/>`-2 = refusal`<br/>`-1 = don't know` <br/> `-7 = not yet coded (temporary)` <br/> `-4 = not codable` | `-2 = refusal`<br/>`-1 = don't know` |
| **Description** |  |  |

```{r assign w1, echo=F}
opal.assign.table.tibble(o, 'SHARE_w1','SHARE.sharew1_rel2-6-0_dn', variables=list('dn005c', 'dn004_'), missings = TRUE)
SHARE_w1 <- opal.execute(o,'SHARE_w1')

load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SHARE/RData/w1/country.RData")

SHARE_w1 <- SHARE_w1 %>% 
  left_join(country, by = "id") 
```

```{r local w1, echo=F}
vbl <- SHARE_w1$dn005c
vbl_miss <- c(-2,-1, -7, -4)
kable(Categorical_summary(vbl, missing_values = vbl_miss)[3], caption = names(SHARE_w1)[3])
kable(Categorical_summary(vbl, missing_values = vbl_miss)[1], caption = "Category")
ggplot(SHARE_w1 %>% filter(!is.na(dn005c),!dn005c %in% vbl_miss), aes(x=factor(dn005c))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency") + scale_x_discrete(breaks = unique((SHARE_w1 %>% filter(dn005c %% 20 == 0))$dn005c)) + theme(axis.text.x = element_text(angle = 45))

vbl <- SHARE_w1$dn004_
vbl_miss <- c(-2,-1)
kable(Categorical_summary(vbl, missing_values = vbl_miss)[3], caption = names(SHARE_w1)[2])
kable(Categorical_summary(vbl, missing_values = vbl_miss)[1], caption = "Category")
ggplot(SHARE_w1, aes(x=factor(dn004_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")
```

#### Harmonisation algorithm

To compute `migration` from the study-specific variable it has to be recoded as follows:


* `dn004_ = 1 into 0`
* `dn004_ = 5:`
    + `if 'country' (harmonized variable) match with 'dn005c' into 999`
    + `if 'country' (HZ) is different as the country code in dn005c into 1`
    + `dn005c = -7 into 1`
    + `dn005c = -2 into 1`
    + `dn005c = -1 into 1`
    + `dn005c = -4 into 1`
* `-2 into 997`
* `-1 into 998`
* `NA into 999`

**R script:**

```{r harmo w1}
SHARE_w1$dn005c_2 <- NA
SHARE_w1$dn005c_2[!(SHARE_w1$dn005c %in% c(40,276,278,280,752,528, 530, 530,724,380,249, 250,208,300,756,56, 376,203,616,372,372,348,620,705,233,191))] <- 1
SHARE_w1$dn005c_2[is.na(SHARE_w1$dn005c)] <- 999
SHARE_w1$dn005c_2[SHARE_w1$dn005c == -2] <- 1
SHARE_w1$dn005c_2[SHARE_w1$dn005c == -1] <- 1
SHARE_w1$dn005c_2[SHARE_w1$dn005c == -7] <- 1
SHARE_w1$dn005c_2[SHARE_w1$dn005c == -4] <- 1
SHARE_w1$dn005c_2[SHARE_w1$dn005c %in% c(40,276,278,280,752,528, 530, 530,724,380,249, 250,208,300,756,56, 376,203,616,372,372,348,620,705,233,191)] <- SHARE_w1$dn005c[SHARE_w1$dn005c %in% c(40,276,278,280,752,528, 530, 530,724,380,249, 250,208,300,756,56, 376,203,616,372,372,348,620,705,233,191)]
SHARE_w1$dn005c_2 <- car::recode(SHARE_w1$dn005c_2, "40=51; 276=59; 278=59;280=59; 752=71; 528=66; 530=66; 530=66; 724=70; 380=63; 249=58; 250=58; 208=55; 300=60; 756=72; 56=52;  376=33; 203=54; 616=65; 372=62; 372=62; 348=61; 620=68; 705=69; 233= 56; 191=53")
SHARE_w1$dn005c_2[!(SHARE_w1$dn005c_2 %in% c(999,1))] <- with(SHARE_w1[!(SHARE_w1$dn005c_2 %in% c(999,1)),], ifelse(dn005c_2 == country, 0,1))

SHARE_ds_w1 <- tibble(id=SHARE_w1$id)
SHARE_ds_w1$migration <- car::recode(SHARE_w1$dn004_, "1=0; 5=1; -2=997; -1=998; NA=999")
SHARE_ds_w1$migration[SHARE_w1$dn004_==5 & SHARE_w1$dn005c_2 == 0] <- 999
```

#### Statistical description of the new harmonised variable

```{r descript w1, echo=F}

vbl <- SHARE_ds_w1$migration
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(SHARE_ds_w1, aes(x=factor(migration))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

```

#### Validation
```{r crosstabulation w1, echo=F}

```



### Wave 2

#### Study-specific variable description

| **Name** | `dn005c` |`dn004_` |
|-|-|-|
| **Label** | `other country of birth-country code` |`country of birth` |
| **Table name** | `sharew2_rel2-6-0_dn` |`sharew2_rel2-6-0_dn` |
| **Categories** | `2 = Africa` <br/>`54 = Afghanistan` <br/> `8 = Albania` <br/> `10 = Antarctica` <br/> `12 = Algeria` <br/> `16 = American Samoa` <br/> `20 = Andorra` <br/> `24 = Angola` <br/> `28 = Antigua and Barbuda` <br/> `31 = Azerbaijan` <br/> `32 = Argentina` <br/> `36 = Australia` <br/> `40 = Austria` <br/> `44 = Bahamas` <br/> `48 = Bahrain`  <br/> `50 = Bangladesh` <br/> `51 = Armenia` <br/> `52 = Barbados` <br/> `56 = Belgium` <br/> `60 = Bermuda` <br/> `64 = Bhutan` <br/> `68 = Bolivia` <br/> `70 = Bosnia and Herzegowina` <br/> `72 = Botswana` <br/> `74 = Bouvet Island` <br/> `76 = Brazil` <br/> `80 = British Antarctic Territory` <br/> `84 = Belize` <br/> `86 = British Indian Ocean Territory` <br/> `90 = Solomon Islands` <br/> `92 = Virgin Islands (british)` <br/> `96 = Brunei Darussalam` <br/> `100 = Bulgaria` <br/> `104 = Myanmar` <br/> `108 = Burundi` <br/> `112 = Belarus` <br/> `116 = Cambodia` <br/> `120 = Cameroon` <br/> `124 = Canada` <br/> `128 = Canton and Enderbury Islands` <br/> `132 = Cape Verde` <br/> `136 = Cayman Islands` <br/> `140 = Central African Republic` <br/> `144 = Sri Lanka` <br/> `148 = Chad` <br/> `152 = Chile ` <br/> `...` | `1 = yes` <br/> `2 = no` |
| **Missings** | `NA`<br/>`-2 = refusal`<br/>`-1 = don't know` <br/> `-7 = not yet coded (temporary)` <br/> `-4 = not codable` | `-2 = refusal`<br/>`-1 = don't know` |
| **Description** |  |  |


```{r assign w2, echo=F}

opal.assign.table.tibble(o, 'SHARE_w2','SHARE.sharew2_rel2-6-0_dn', variables=list('dn005c', 'dn004_'), missings = TRUE)
SHARE_w2 <- opal.execute(o,'SHARE_w2')

load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SHARE/RData/w2/country.RData")

SHARE_w2 <- SHARE_w2 %>% 
  left_join(country, by = "id") %>% 
  left_join(SHARE_ds_w1, by = "id")
```

```{r local w2, echo=F}

vbl <- SHARE_w2$dn005c
vbl_miss <- c(-2,-1, -4)
kable(Categorical_summary(vbl, missing_values = vbl_miss)[3], caption = names(SHARE_w2)[3])
kable(Categorical_summary(vbl, missing_values = vbl_miss)[1], caption = "Category")
ggplot(SHARE_w2 %>% filter(!is.na(dn005c),!dn005c %in% vbl_miss), aes(x=factor(dn005c))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency") + scale_x_discrete(breaks = unique((SHARE_w2 %>% filter(dn005c %% 20 == 0))$dn005c)) + theme(axis.text.x = element_text(angle = 45))

vbl <- SHARE_w2$dn004_
vbl_miss <- c(-2,-1)
kable(Categorical_summary(vbl, missing_values = vbl_miss)[3], caption = names(SHARE_w2)[2])
kable(Categorical_summary(vbl, missing_values = vbl_miss)[1], caption = "Category")
ggplot(SHARE_w2, aes(x=factor(dn004_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```

#### Harmonisation algorithm

To compute `migration` from the study-specific variable it has to be recoded as follows:


* `dn004_ = 1 into 0`
* `dn004_ = 5:`
    + `if 'country' (harmonized variable) match with 'dn005c' into 999`
    + `if 'country' (HZ) is different as the country code in dn005c into 1`
    + `dn005c = -7 into 1`
    + `dn005c = -2 into 1`
    + `dn005c = -1 into 1`
    + `dn005c = -4 into 1`
* `-2 into 997`
* `-1 into 998`
* `NA into 999`


**R script:**

```{r harmo w2}
SHARE_w2$dn005c_2 <- NA
SHARE_w2$dn005c_2[!(SHARE_w2$dn005c %in% c(40,276,278,280,752,528, 530, 530,724,380,249, 250,208,300,756,56, 376,203,616,372,372,348,620,705,233,191))] <- 1
SHARE_w2$dn005c_2[is.na(SHARE_w2$dn005c)] <- 999
SHARE_w2$dn005c_2[SHARE_w2$dn005c == -2] <- 1
SHARE_w2$dn005c_2[SHARE_w2$dn005c == -1] <- 1
SHARE_w2$dn005c_2[SHARE_w2$dn005c == -7] <- 1
SHARE_w2$dn005c_2[SHARE_w2$dn005c == -4] <- 1
SHARE_w2$dn005c_2[SHARE_w2$dn005c %in% c(40,276,278,280,752,528, 530, 530,724,380,249, 250,208,300,756,56, 376,203,616,372,372,348,620,705,233,191)] <- SHARE_w2$dn005c[SHARE_w2$dn005c %in% c(40,276,278,280,752,528, 530, 530,724,380,249, 250,208,300,756,56, 376,203,616,372,372,348,620,705,233,191)]
SHARE_w2$dn005c_2 <- car::recode(SHARE_w2$dn005c_2, "40=51; 276=59; 278=59;280=59; 752=71; 528=66; 530=66; 530=66; 724=70; 380=63; 249=58; 250=58; 208=55; 300=60; 756=72; 56=52;  376=33; 203=54; 616=65; 372=62; 372=62; 348=61; 620=68; 705=69; 233= 56; 191=53")
SHARE_w2$dn005c_2[!(SHARE_w2$dn005c_2 %in% c(999,1))] <- with(SHARE_w2[!(SHARE_w2$dn005c_2 %in% c(999,1)),], ifelse(dn005c_2 == country, 0,1))


SHARE_ds_w2 <- tibble(id=SHARE_w2$id)
SHARE_ds_w2$migration <- car::recode(SHARE_w2$dn004_, "1=0; 5=1; -2=997; -1=998; NA=999")
SHARE_ds_w2$migration[SHARE_w2$dn004_==5 & SHARE_w2$dn005c_2 == 0] <- 999
SHARE_ds_w2$migration[SHARE_ds_w2$migration == 999 & !is.na(SHARE_w2$migration)] <- SHARE_w2$migration[SHARE_ds_w2$migration == 999 & !is.na(SHARE_w2$migration)]

```

#### Statistical description of the new harmonised variable
```{r descript w2, echo=F}

vbl <- SHARE_ds_w2$migration
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(SHARE_ds_w2, aes(x=factor(migration))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

```

#### Validation
```{r crosstabulation w2, echo=F}

```

### Wave 3

Not found.


### Wave 4

#### Study-specific variable description

| **Name** | `dn005c` | `dn004_` |
|-|-|-|
| **Label** | `other country of birth-country code` |`country of birth` |
| **Table name** | `sharew4_rel1-1-1_dn` |`sharew4_rel1-1-1_dn` |
| **Categories** | `2 = Africa` <br/>`54 = Afghanistan` <br/> `8 = Albania` <br/> `10 = Antarctica` <br/> `12 = Algeria` <br/> `16 = American Samoa` <br/> `20 = Andorra` <br/> `24 = Angola` <br/> `28 = Antigua and Barbuda` <br/> `31 = Azerbaijan` <br/> `32 = Argentina` <br/> `36 = Australia` <br/> `40 = Austria` <br/> `44 = Bahamas` <br/> `48 = Bahrain`  <br/> `50 = Bangladesh` <br/> `51 = Armenia` <br/> `52 = Barbados` <br/> `56 = Belgium` <br/> `60 = Bermuda` <br/> `64 = Bhutan` <br/> `68 = Bolivia` <br/> `70 = Bosnia and Herzegowina` <br/> `72 = Botswana` <br/> `74 = Bouvet Island` <br/> `76 = Brazil` <br/> `80 = British Antarctic Territory` <br/> `84 = Belize` <br/> `86 = British Indian Ocean Territory` <br/> `90 = Solomon Islands` <br/> `92 = Virgin Islands (british)` <br/> `96 = Brunei Darussalam` <br/> `100 = Bulgaria` <br/> `104 = Myanmar` <br/> `108 = Burundi` <br/> `112 = Belarus` <br/> `116 = Cambodia` <br/> `120 = Cameroon` <br/> `124 = Canada` <br/> `128 = Canton and Enderbury Islands` <br/> `132 = Cape Verde` <br/> `136 = Cayman Islands` <br/> `140 = Central African Republic` <br/> `144 = Sri Lanka` <br/> `148 = Chad` <br/> `152 = Chile ` <br/> `...` | `1 = yes` <br/> `2 = no` |
| **Missings** | `NA`<br/>`-2 = refusal`<br/>`-1 = don't know` <br/> `-7 = not yet coded (temporary)` <br/> `-4 = not codable` | `-2 = refusal`<br/>`-1 = don't know` |
| **Description** |  |  |

```{r assign w4, echo=F}

opal.assign.table.tibble(o, 'SHARE_w4','SHARE.sharew4_rel1-1-1_dn', variables=list('dn005c', 'dn004_'), missings = TRUE)
SHARE_w4 <- opal.execute(o,'SHARE_w4')

load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SHARE/RData/w4/country.RData")

SHARE_w4 <- SHARE_w4 %>% 
  left_join(country, by = "id") %>% 
  left_join(SHARE_ds_w2, by = "id")

```

```{r local w4, echo=F}

vbl <- SHARE_w4$dn005c
vbl_miss <- c(-2,-1, -4)
kable(Categorical_summary(vbl, missing_values = vbl_miss)[3], caption = names(SHARE_w4)[3])
kable(Categorical_summary(vbl, missing_values = vbl_miss)[1], caption = "Category")
ggplot(SHARE_w4 %>% filter(!is.na(dn005c),!dn005c %in% vbl_miss), aes(x=factor(dn005c))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency") + scale_x_discrete(breaks = unique((SHARE_w4 %>% filter(dn005c %% 20 == 0))$dn005c)) + theme(axis.text.x = element_text(angle = 45))

vbl <- SHARE_w4$dn004_
vbl_miss <- c(-2,-1)
kable(Categorical_summary(vbl, missing_values = vbl_miss)[3], caption = names(SHARE_w4)[2])
kable(Categorical_summary(vbl, missing_values = vbl_miss)[1], caption = "Category")
ggplot(SHARE_w4, aes(x=factor(dn004_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")
```

#### Harmonisation algorithm

To compute `migration` from the study-specific variable it has to be recoded as follows:


* `dn004_ = 1 into 0`
* `dn004_ = 5:`
    + `if 'country' (harmonized variable) match with 'dn005c' into 999`
    + `if 'country' (HZ) is different as the country code in dn005c into 1`
    + `dn005c = -7 into 1`
    + `dn005c = -2 into 1`
    + `dn005c = -1 into 1`
    + `dn005c = -4 into 1`
* `-2 into 997`
* `-1 into 998`
* `NA into 999`


**R script:**

```{r harmo w4}
SHARE_w4$dn005c_2 <- NA
SHARE_w4$dn005c_2[!(SHARE_w4$dn005c %in% c(40,276,278,280,752,528, 530, 530,724,380,249, 250,208,300,756,56, 376,203,616,372,372,348,620,705,233,191))] <- 1
SHARE_w4$dn005c_2[is.na(SHARE_w4$dn005c)] <- 999
SHARE_w4$dn005c_2[SHARE_w4$dn005c == -2] <- 1
SHARE_w4$dn005c_2[SHARE_w4$dn005c == -1] <- 1
SHARE_w4$dn005c_2[SHARE_w4$dn005c == -7] <- 1
SHARE_w4$dn005c_2[SHARE_w4$dn005c == -4] <- 1
SHARE_w4$dn005c_2[SHARE_w4$dn005c %in% c(40,276,278,280,752,528, 530, 530,724,380,249, 250,208,300,756,56, 376,203,616,372,372,348,620,705,233,191)] <- SHARE_w4$dn005c[SHARE_w4$dn005c %in% c(40,276,278,280,752,528, 530, 530,724,380,249, 250,208,300,756,56, 376,203,616,372,372,348,620,705,233,191)]
SHARE_w4$dn005c_2 <- car::recode(SHARE_w4$dn005c_2, "40=51; 276=59; 278=59;280=59; 752=71; 528=66; 530=66; 530=66; 724=70; 380=63; 249=58; 250=58; 208=55; 300=60; 756=72; 56=52;  376=33; 203=54; 616=65; 372=62; 372=62; 348=61; 620=68; 705=69; 233= 56; 191=53")
SHARE_w4$dn005c_2[!(SHARE_w4$dn005c_2 %in% c(999,1))] <- with(SHARE_w4[!(SHARE_w4$dn005c_2 %in% c(999,1)),], ifelse(dn005c_2 == country, 0,1))

SHARE_ds_w4 <- tibble(id=SHARE_w4$id)
SHARE_ds_w4$migration <- car::recode(SHARE_w4$dn004_, "1=0; 5=1; -2=997; -1=998; NA=999")
SHARE_ds_w4$migration[SHARE_w4$dn004_==5 & SHARE_w4$dn005c_2 == 0] <- 999
SHARE_ds_w4$migration[SHARE_ds_w4$migration == 999 & !is.na(SHARE_w4$migration)] <- SHARE_w4$migration[SHARE_ds_w4$migration == 999 & !is.na(SHARE_w4$migration)]
```

#### Statistical description of the new harmonised variable
```{r descript w4, echo=F}

vbl <- SHARE_ds_w4$migration
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(SHARE_ds_w4, aes(x=factor(migration))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

```

#### Validation
```{r crosstabulation w4, echo=F}

```


### Wave 5

#### Study-specific variable description

| **Name** | `dn005c` | `dn004_` |
|-|-|-|
| **Label** | `other country of birth-country code` |`country of birth` |
| **Table name** | `sharew5_rel1-0-0_dn` | `sharew5_rel1-0-0_dn` |
| **Categories** | `2 = Africa` <br/>`54 = Afghanistan` <br/> `8 = Albania` <br/> `10 = Antarctica` <br/> `12 = Algeria` <br/> `16 = American Samoa` <br/> `20 = Andorra` <br/> `24 = Angola` <br/> `28 = Antigua and Barbuda` <br/> `31 = Azerbaijan` <br/> `32 = Argentina` <br/> `36 = Australia` <br/> `40 = Austria` <br/> `44 = Bahamas` <br/> `48 = Bahrain`  <br/> `50 = Bangladesh` <br/> `51 = Armenia` <br/> `52 = Barbados` <br/> `56 = Belgium` <br/> `60 = Bermuda` <br/> `64 = Bhutan` <br/> `68 = Bolivia` <br/> `70 = Bosnia and Herzegowina` <br/> `72 = Botswana` <br/> `74 = Bouvet Island` <br/> `76 = Brazil` <br/> `80 = British Antarctic Territory` <br/> `84 = Belize` <br/> `86 = British Indian Ocean Territory` <br/> `90 = Solomon Islands` <br/> `92 = Virgin Islands (british)` <br/> `96 = Brunei Darussalam` <br/> `100 = Bulgaria` <br/> `104 = Myanmar` <br/> `108 = Burundi` <br/> `112 = Belarus` <br/> `116 = Cambodia` <br/> `120 = Cameroon` <br/> `124 = Canada` <br/> `128 = Canton and Enderbury Islands` <br/> `132 = Cape Verde` <br/> `136 = Cayman Islands` <br/> `140 = Central African Republic` <br/> `144 = Sri Lanka` <br/> `148 = Chad` <br/> `152 = Chile ` <br/> `...` | `1 = yes` <br/> `2 = no` |
| **Missings** | `NA`<br/>`-2 = refusal`<br/>`-1 = don't know` <br/> `-7 = not yet coded (temporary)` <br/> `-4 = not codable` | `-2 = refusal`<br/>`-1 = don't know` |
| **Description** |  |  |


```{r assign w5, echo=F}

opal.assign.table.tibble(o, 'SHARE_w5','SHARE.sharew5_rel1-0-0_dn', variables=list('dn005c', 'dn004_'), missings = TRUE)
SHARE_w5 <- opal.execute(o,'SHARE_w5')

load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/SHARE/RData/w2/country.RData")

SHARE_w5 <- SHARE_w5 %>% 
  left_join(country, by = "id") %>% 
  left_join(SHARE_ds_w4, by = "id")
```

```{r local w5, echo=F}

vbl <- SHARE_w5$dn005c
vbl_miss <- c(-2,-1, -4)
kable(Categorical_summary(vbl, missing_values = vbl_miss)[3], caption = names(SHARE_w5)[3])
kable(Categorical_summary(vbl, missing_values = vbl_miss)[1], caption = "Category")
ggplot(SHARE_w5 %>% filter(!is.na(dn005c),!dn005c %in% vbl_miss), aes(x=factor(dn005c))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency") + scale_x_discrete(breaks = unique((SHARE_w5 %>% filter(dn005c %% 20 == 0))$dn005c)) + theme(axis.text.x = element_text(angle = 45))

vbl <- SHARE_w5$dn004_
vbl_miss <- c(-2,-1)
kable(Categorical_summary(vbl, missing_values = vbl_miss)[3], caption = names(SHARE_w5)[2])
kable(Categorical_summary(vbl, missing_values = vbl_miss)[1], caption = "Category")
ggplot(SHARE_w5, aes(x=factor(dn004_))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")
```

#### Harmonisation algorithm

To compute `migration` from the study-specific variable it has to be recoded as follows:


* `dn004_ = 1 into 0`
* `dn004_ = 5:`
    + `if 'country' (harmonized variable) match with 'dn005c' into 999`
    + `if 'country' (HZ) is different as the country code in dn005c into 1`
    + `dn005c = -7 into 1`
    + `dn005c = -2 into 1`
    + `dn005c = -1 into 1`
    + `dn005c = -4 into 1`
* `-2 into 997`
* `-1 into 998`
* `NA into 999`


**R script:**

```{r harmo w5}
SHARE_w5$dn005c_2 <- NA
SHARE_w5$dn005c_2[!(SHARE_w5$dn005c %in% c(40,276,278,280,752,528, 530, 530,724,380,249, 250,208,300,756,56, 376,203,616,372,372,348,620,705,233,191))] <- 1
SHARE_w5$dn005c_2[is.na(SHARE_w5$dn005c)] <- 999
SHARE_w5$dn005c_2[SHARE_w5$dn005c == -2] <- 1
SHARE_w5$dn005c_2[SHARE_w5$dn005c == -1] <- 1
SHARE_w5$dn005c_2[SHARE_w5$dn005c == -7] <- 1
SHARE_w5$dn005c_2[SHARE_w5$dn005c == -4] <- 1
SHARE_w5$dn005c_2[SHARE_w5$dn005c %in% c(40,276,278,280,752,528, 530, 530,724,380,249, 250,208,300,756,56, 376,203,616,372,372,348,620,705,233,191)] <- SHARE_w5$dn005c[SHARE_w5$dn005c %in% c(40,276,278,280,752,528, 530, 530,724,380,249, 250,208,300,756,56, 376,203,616,372,372,348,620,705,233,191)]
SHARE_w5$dn005c_2 <- car::recode(SHARE_w5$dn005c_2, "40=51; 276=59; 278=59;280=59; 752=71; 528=66; 530=66; 530=66; 724=70; 380=63; 249=58; 250=58; 208=55; 300=60; 756=72; 56=52;  376=33; 203=54; 616=65; 372=62; 372=62; 348=61; 620=68; 705=69; 233= 56; 191=53")
SHARE_w5$dn005c_2[!(SHARE_w5$dn005c_2 %in% c(999,1))] <- with(SHARE_w5[!(SHARE_w5$dn005c_2 %in% c(999,1)),], ifelse(dn005c_2 == country, 0,1))

SHARE_ds_w5 <- tibble(id=SHARE_w5$id)
SHARE_ds_w5$migration <- car::recode(SHARE_w5$dn004_, "1=0; 5=1; -2=997; -1=998; NA=999")
SHARE_ds_w5$migration[SHARE_w5$dn004_==5 & SHARE_w5$dn005c_2 == 0] <- 999
SHARE_ds_w5$migration[SHARE_ds_w5$migration == 999 & !is.na(SHARE_w5$migration)] <- SHARE_w5$migration[SHARE_ds_w5$migration == 999 & !is.na(SHARE_w5$migration)]
```

#### Statistical description of the new harmonised variable
```{r descript w5, echo=F}

vbl <- SHARE_ds_w5$migration
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(SHARE_ds_w5, aes(x=factor(migration))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

```

#### Validation
```{r crosstabulation w5, echo=F}

```

```{r helpfunctions, echo=F}


labelling <- function(l.hds,m.hds){
  
  # Labelling of the tibbles with categorical data and creating new tibbles with all missings recodified as NA

  for(name in names(l.hds)) {
    # In the aux_object we copy the old tibble to recodify all missing values as NA.
    aux_object <- l.hds[[name]]
    # Labelling of variables
    label(l.hds[[name]][[2]]) <- label(aux_object[[2]]) <- ds_label
    # Labelling of categories (for continues variables, only missing values)
    l.hds[[name]][[2]] <- labelled(l.hds[[name]][[2]], labels = cat_label)
    aux_object[[2]] <- car::recode(aux_object[[2]], "miss_values_vector=NA")
    # Labelling of categories (for categorical variables, only non-missing values)
    aux_object[[2]] <- labelled(aux_object[[2]], labels = cat_label[1:2])
  # Saving the recodified tibble in list m.hds
    m.hds[[name]] <- aux_object
    rm(aux_object)
  }
  return(list(l.hds,m.hds))

}

# Creation of summary tables for categorical data

summaries <- function(l.hds,m.hds,lnames){

  # Creation of columns with categories and labels
  t.hds <- frq(l.hds[[1]][2])[[1]][,c(1,2)] 
  # For each wave/population in l.hds, add the correponding values
  for (i in seq_along(l.hds)){
    t.hds[2+i] <- frq(l.hds[[i]][2])[[1]][,4] 
  }
  # Add sample size for each wave/population
  t.hds[13,] <- c("n", "sample size", sapply(l.hds,function(wave) length(wave[[1]]))
  )
  # Add wave/population names
  names(t.hds)<-c("val", "label",lnames)
  return(t.hds)
  
}

# Creation of trajectories table for each population

trajectories <- function(m.hds,vbl_name){
  
  # First wave data
  dbb <- m.hds[[1]][,c("id",vbl_name)]
  # Merge with next waves data
  for(ind in 2:length(m.hds)){
    dbb <- merge(dbb, m.hds[[ind]][,c("id",vbl_name)], by = "id", suffixes = c("", paste0(".",names(m.hds)[ind])), all = T)
  }
  names(dbb) <- c("id", names(m.hds))
  
  # Glue individual data through all waves into trajectories
  v.dbb <- dbb[,2]
  for(ind in 2:length(m.hds)){
    v.dbb <- paste(v.dbb,dbb[,ind+1],sep="")
  }
  # Trajectories and frequencies
  f.dbb <- frq(v.dbb)[[1]][,c(1,2,4)]
  return(f.dbb)
  
}

# Save data tables

savingRD <- function(l.hds,vbl_name){
  
  for(index in seq_along(l.hds)){
    assign(vbl_name,l.hds[[index]])
    save(vbl_name,list = vbl_name, file = paste0(datafolder,names(l.hds)[index],"/",vbl_name,".RData"))
  }

}

```


#### Corrections

There are 12 individuals that answered in one wave that they are migrants and in other that they are not. Those inviduals have been recoded as missing values (999):

```{r correction, echo=F}
share.cw <- list(w1 = c("w1","W1"), w2 = c("w2","W2"), w3 = c("w3","W3"), w4 = c("w4","W4"), w5 = c("w5","W5"))

# Consider only harmonised waves
l.hds <- lapply(share.cw, function(wname) if(exists(paste0("SHARE_ds_",wname[1]))){wname = list(get(paste0("SHARE_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

lmlist <- labelling(lmlist[[1]],lmlist[[2]])
m.hds <- lmlist[[2]]
vbl_name <- hd_vbl
# First wave data
dbb <- m.hds[[1]][,c("id",vbl_name)]
  # Merge with next waves data
for(ind in 2:length(m.hds)){
  dbb <- merge(dbb, m.hds[[ind]][,c("id",vbl_name)], by = "id", suffixes = c("", paste0(".",names(m.hds)[ind])), all = T)
  }
names(dbb) <- c("id", names(m.hds))

iddiff_1 <- dbb$id[which(dbb$w1==0 & dbb$w2 == 1)]
iddiff_2 <- dbb$id[which(dbb$w1==1 & dbb$w2 == 0)]
iddiff_3 <- dbb$id[which(dbb$w1==1 & dbb$w4 == 0)]
iddiff_4 <- dbb$id[which(dbb$w1==1 & dbb$w5 == 0)]
iddiff_5 <- dbb$id[which(dbb$w2==1 & dbb$w4 == 0)]
iddiff_6 <- dbb$id[which(dbb$w4==1 & dbb$w5 == 0)]
iddiff_7 <- dbb$id[which(dbb$w2==1 & dbb$w5 == 0)]

iddiff <- c(iddiff_1, iddiff_2, iddiff_3, iddiff_4, iddiff_5, iddiff_6,iddiff_7)
SHARE_ds_w1$migration[which(SHARE_ds_w1$id %in% c(iddiff))] <- 999
SHARE_ds_w2$migration[which(SHARE_ds_w2$id %in% c(iddiff))] <- 999
SHARE_ds_w4$migration[which(SHARE_ds_w4$id %in% c(iddiff))] <- 999
SHARE_ds_w5$migration[which(SHARE_ds_w5$id %in% c(iddiff))] <- 999
```




## Summary of descriptive statistics of the harmonised variable accross populations and waves

Two tables are generated: 

1. Percentages of categories in each harmonised variable.

2. Frequencies and percentages of individuals with different trajectories.






```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names



share.cw <- list(w1 = c("w1","W1"), w2 = c("w2","W2"), w3 = c("w3","W3"), w4 = c("w4","W4"), w5 = c("w5","W5"))

# Consider only harmonised waves
l.hds <- lapply(share.cw, function(wname) if(exists(paste0("SHARE_ds_",wname[1]))){wname = list(get(paste0("SHARE_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing table of categories
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Printing table of trajectories
  f.dbb <- trajectories(lmlist[[2]],vbl_name=hd_vbl)
  print(knitr::kable(f.dbb))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```


 







# Quality estimation



<!-- ########################################################## -->
<!--- # Close OPAL R Session -->
```{r closeRsession, echo=FALSE}
opal.logout(o)
```
